move all lumi functions to /lib/lumi.ts
