+++ src/lib/lumi.ts
@@ -5,13 +5,25 @@ import {
     assertNumber,
     getFromLookup,
     getKey,
+    createLogger,
+    printNumbersAsHexSequence,
+    toNumber,
+    assertEndpoint,
+    assertString,
+    hasAlreadyProcessedMessage,
+    isLegacyEnabled,
+    noOccupancySince,
+    isObject,
+    isString,
+    getOptions,
+    assertObject,
+    calibrateAndPrecisionRoundOptions,
 } from './utils';
 
 import * as ota from './ota';
 import fz from '../converters/fromZigbee';
-import tz from '../converters/toZigbee';
 import * as globalStore from './store';
-import {Fz, Definition, KeyValue, KeyValueAny, Tz, ModernExtend, Range} from './types';
+import {Fz, Definition, KeyValue, KeyValueAny, Tz, ModernExtend, Range, KeyValueNumberString} from './types';
 import * as modernExtend from './modernExtend';
 import * as exposes from './exposes';
 
@@ -781,7 +793,6 @@ export const numericAttributes2Payload = async (msg: Fz.Message, meta: Fz.Meta, 
     return payload;
 };
 
-// For RTCZCGQ11LM
 type LumiPresenceRegionZone = {x: number, y: number}
 
 const lumiPresenceConstants = {
@@ -1248,6 +1259,9 @@ export const trv = {
 };
 
 export const manufacturerCode = 0x115f;
+const manufacturerOptions = {
+    lumi: {manufacturerCode: manufacturerCode, disableDefaultResponse: true},
+};
 
 export const lumiModernExtend = {
     lumiLight: (args?: Omit<modernExtend.LightArgs, 'colorTemp'> & {colorTemp?: true, powerOutageMemory?: 'switch' | 'light'}) => {
@@ -1255,16 +1269,16 @@ export const lumiModernExtend = {
         const colorTemp: {range: Range, startup: boolean} = args.colorTemp ? {startup: false, range: [153, 370]} : undefined;
         const result = modernExtend.light({effect: false, powerOnBehavior: false, ...args, colorTemp});
         result.fromZigbee.push(
-            fz.lumi_bulb_interval, fz.ignore_occupancy_report, fz.ignore_humidity_report,
+            fromZigbee.lumi_bulb_interval, fz.ignore_occupancy_report, fz.ignore_humidity_report,
             fz.ignore_pressure_report, fz.ignore_temperature_report, fromZigbee.lumi_specific,
         );
         result.exposes.push(e.device_temperature(), e.power_outage_count());
 
         if (args.powerOutageMemory === 'switch') {
-            result.toZigbee.push(tz.lumi_switch_power_outage_memory);
+            result.toZigbee.push(toZigbee.lumi_switch_power_outage_memory);
             result.exposes.push(e.power_outage_memory());
         } else if (args.powerOutageMemory === 'light') {
-            result.toZigbee.push(tz.lumi_light_power_outage_memory);
+            result.toZigbee.push(toZigbee.lumi_light_power_outage_memory);
             result.exposes.push(e.power_outage_memory().withAccess(ea.STATE_SET));
         }
 
@@ -1475,7 +1489,7 @@ const feederDaysLookup = {
 };
 
 export const fromZigbee = {
-    aqara_feeder: {
+    lumi_feeder: {
         cluster: 'manuSpecificLumi',
         type: ['attributeReport', 'readResponse'],
         convert: (model, msg, publish, options, meta) => {
@@ -1592,10 +1606,1221 @@ export const fromZigbee = {
             return await numericAttributes2Payload(msg, meta, model, options, msg.data);
         },
     } satisfies Fz.Converter,
+    lumi_co2: {
+        cluster: 'msCO2',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            return {co2: Math.floor(msg.data.measuredValue)};
+        },
+    } satisfies Fz.Converter,
+    lumi_pm25: {
+        cluster: 'pm25Measurement',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data['measuredValue']) {
+                return {pm25: msg.data['measuredValue'] / 1000};
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_trv: {
+        cluster: 'manuSpecificLumi',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValue = {};
+            Object.entries(msg.data).forEach(([key, value]) => {
+                switch (parseInt(key)) {
+                case 0x0271:
+                    result['system_mode'] = getFromLookup(value, {1: 'heat', 0: 'off'});
+                    break;
+                case 0x0272:
+                    // @ts-expect-error
+                    Object.assign(result, trv.decodePreset(value));
+                    break;
+                case 0x0273:
+                    result['window_detection'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x0274:
+                    result['valve_detection'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x0277:
+                    result['child_lock'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x0279:
+                    assertNumber(value);
+                    result['away_preset_temperature'] = (value / 100).toFixed(1);
+                    break;
+                case 0x027b:
+                    result['calibrated'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x027e:
+                    result['sensor'] = getFromLookup(value, {1: 'external', 0: 'internal'});
+                    break;
+                case 0x040a:
+                    result['battery'] = value;
+                    break;
+                case 0x027a:
+                    result['window_open'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x0275:
+                    result['valve_alarm'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 247: {
+                    // @ts-expect-error
+                    const heartbeat = trv.decodeHeartbeat(meta, model, value);
+
+                    meta.logger.debug(`${model.model}: Processed heartbeat message into payload ${JSON.stringify(heartbeat)}`);
+
+                    if (heartbeat.firmware_version) {
+                        // Overwrite the "placeholder" version `0.0.0_0025` advertised by `genBasic`
+                        // with the correct version from the heartbeat.
+                        // This is not reflected in the frontend unless the device is reconfigured
+                        // or the whole service restarted.
+                        // See https://github.com/Koenkk/zigbee-herdsman-converters/pull/5363#discussion_r1081477047
+                        // @ts-expect-error
+                        meta.device.softwareBuildID = heartbeat.firmware_version;
+                        delete heartbeat.firmware_version;
+                    }
+
+                    Object.assign(result, heartbeat);
+                    break;
+                }
+                case 0x027d:
+                    result['schedule'] = getFromLookup(value, {1: true, 0: false});
+                    break;
+                case 0x0276: {
+                    // @ts-expect-error
+                    const schedule = trv.decodeSchedule(value);
+                    result['schedule_settings'] = trv.stringifySchedule(schedule);
+                    break;
+                }
+                case 0x00EE: {
+                    meta.device.meta.aqaraFileVersion = value;
+                    meta.device.save();
+                    break;
+                }
+                case 0xfff2:
+                case 0x00ff: // 4e:27:49:bb:24:b6:30:dd:74:de:53:76:89:44:c4:81
+                case 0x027c: // 0x00
+                case 0x0280: // 0x00/0x01
+                    meta.logger.debug(`zigbee-herdsman-converters:lumi_trv: Unhandled key ${key} = ${value}`);
+                    break;
+                default:
+                    meta.logger.warn(`zigbee-herdsman-converters:lumi_trv: Unknown key ${key} = ${value}`);
+                }
+            });
+            return result;
+        },
+    } satisfies Fz.Converter,
+    lumi_presence_region_events: {
+        cluster: 'manuSpecificLumi',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const payload: KeyValue = {};
+            const log = createLogger(meta.logger, 'lumi', 'lumi_presence');
+
+            Object.entries(msg.data).forEach(([key, value]) => {
+                const eventKey = parseInt(key);
+
+                switch (eventKey) {
+                case presence.constants.region_event_key: {
+                    if (
+                        !Buffer.isBuffer(value) ||
+                        !(typeof value[0] === 'string' || typeof value[0] === 'number') ||
+                        !(typeof value[1] === 'string' || typeof value[1] === 'number')
+                    ) {
+                        log('warn', `action: Unrecognized payload structure '${JSON.stringify(value)}'`);
+                        break;
+                    }
+
+                    const [regionIdRaw, eventTypeCodeRaw] = value;
+                    // @ts-expect-error
+                    const regionId = parseInt(regionIdRaw, 10);
+                    // @ts-expect-error
+                    const eventTypeCode = parseInt(eventTypeCodeRaw, 10);
+
+                    if (Number.isNaN(regionId)) {
+                        log('warn', `action: Invalid regionId "${regionIdRaw}"`);
+                        break;
+                    }
+                    if (!Object.values(presence.constants.region_event_types).includes(eventTypeCode)) {
+                        log('warn', `action: Unknown region event type "${eventTypeCode}"`);
+                        break;
+                    }
+
+                    const eventTypeName = presence.mappers.lumi_presence.region_event_type_names[eventTypeCode];
+                    log('debug', `action: Triggered event (region "${regionId}", type "${eventTypeName}")`);
+                    payload.action = `region_${regionId}_${eventTypeName}`;
+                    break;
+                }
+                }
+            });
+
+            return payload;
+        },
+    } satisfies Fz.Converter,
+    lumi_cube_t1_action_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const value = msg.data['presentValue'];
+            let payload;
+
+            if (value === 0) payload = {action: 'shake'};
+            else if (value === 1) payload = {action: 'throw'};
+            else if (value === 2) payload = {action: '1_min_inactivity'};
+            else if (value === 4) payload = {action: 'hold'};
+            else if (value >= 1024) payload = {action: 'flip_to_side', side: value - 1023};
+            else if (value >= 512) payload = {action: 'tap', side: value - 511};
+            else if (value >= 256) payload = {action: 'slide', side: value - 255};
+            else if (value >= 128) {
+                payload = {
+                    action: 'flip180', side: value - 127,
+                    action_from_side: 7 - value + 127,
+                };
+            } else if (value >= 64) {
+                payload = {
+                    action: 'flip90', side: value % 8 + 1,
+                    action_from_side: Math.floor((value - 64) / 8) + 1,
+                };
+            } else {
+                meta.logger.debug(`${model.model}: unknown action with value ${value}`);
+            }
+            return payload;
+        },
+    } satisfies Fz.Converter,
+    lumi_cube_t1_action_analog: {
+        cluster: 'genAnalogInput',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const value = msg.data['presentValue'];
+            return {
+                action: value < 0 ? 'rotate_left' : 'rotate_right',
+                action_angle: Math.floor(value * 100) / 100,
+            };
+        },
+    } satisfies Fz.Converter,
+    lumi_contact: {
+        cluster: 'genOnOff',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            return {contact: msg.data['onOff'] === 0};
+        },
+    } satisfies Fz.Converter,
+    lumi_power: {
+        cluster: 'genAnalogInput',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            return {power: msg.data['presentValue']};
+        },
+    } satisfies Fz.Converter,
+    lumi_action_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport'],
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            let actionLookup: KeyValueAny = {0: 'hold', 1: 'single', 2: 'double', 3: 'triple', 255: 'release'};
+            if (model.model === 'WXKG12LM') {
+                actionLookup = {...actionLookup, 16: 'hold', 17: 'release', 18: 'shake'};
+            }
+
+            let buttonLookup: KeyValueNumberString = null;
+            if (['WXKG02LM_rev2', 'WXKG07LM', 'WXKG15LM', 'WXKG17LM', 'WXKG22LM'].includes(model.model)) {
+                buttonLookup = {1: 'left', 2: 'right', 3: 'both'};
+            }
+            if (['QBKG12LM', 'QBKG24LM'].includes(model.model)) buttonLookup = {5: 'left', 6: 'right', 7: 'both'};
+            if (['QBKG39LM', 'QBKG41LM', 'WS-EUK02', 'WS-EUK04', 'QBKG20LM', 'QBKG28LM', 'QBKG31LM', 'ZNQBKG25LM'].includes(model.model)) {
+                buttonLookup = {41: 'left', 42: 'right', 51: 'both'};
+            }
+            if (['QBKG25LM', 'QBKG26LM', 'QBKG29LM', 'QBKG32LM', 'QBKG34LM', 'ZNQBKG31LM', 'ZNQBKG26LM'].includes(model.model)) {
+                buttonLookup = {
+                    41: 'left', 42: 'center', 43: 'right',
+                    51: 'left_center', 52: 'left_right', 53: 'center_right',
+                    61: 'all',
+                };
+            }
+            if (['WS-USC02', 'WS-USC04'].includes(model.model)) {
+                buttonLookup = {41: 'top', 42: 'bottom', 51: 'both'};
+            }
+
+            const action = actionLookup[msg.data['presentValue']];
+            if (buttonLookup) {
+                const button = buttonLookup[msg.endpoint.ID];
+                if (button) {
+                    return {action: `${action}_${button}`};
+                }
+            } else {
+                return {action};
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_lock_report: {
+        cluster: 'genBasic',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data['65328']) {
+                const data = msg.data['65328'];
+                const state = data.substr(2, 2);
+                const action = data.substr(4, 2);
+                const keynum = data.substr(6, 2);
+                if (state == 11) {
+                    if (action == 1) {
+                        // unknown key
+                        return {keyerror: true, inserted: 'unknown'};
+                    }
+                    if (action == 3) {
+                        // explicitly disabled key (i.e. reported lost)
+                        return {keyerror: true, inserted: keynum};
+                    }
+                    if (action == 7) {
+                        // strange object introduced into the cylinder (e.g. a lock pick)
+                        return {keyerror: true, inserted: 'strange'};
+                    }
+                }
+                if (state == 12) {
+                    if (action == 1) {
+                        return {inserted: keynum};
+                    }
+                    if (action == 11) {
+                        return {forgotten: keynum};
+                    }
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_action: {
+        cluster: 'genOnOff',
+        type: ['attributeReport'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (['QBKG04LM', 'QBKG11LM', 'QBKG21LM', 'QBKG03LM', 'QBKG12LM', 'QBKG22LM'].includes(model.model) && msg.data['61440']) {
+                return;
+            }
+
+            if (['QBKG21LM', 'QBKG04LM'].includes(model.model) && msg.endpoint.ID !== 4) return;
+
+            let mapping: KeyValueNumberString = null;
+            if (['QBKG03LM', 'QBKG12LM', 'QBKG22LM'].includes(model.model)) mapping = {4: 'left', 5: 'right', 6: 'both'};
+            if (['WXKG02LM_rev1', 'WXKG02LM_rev2', 'WXKG07LM'].includes(model.model)) mapping = {1: 'left', 2: 'right', 3: 'both'};
+
+            // Maybe other QKBG also support release/hold?
+            const actionLookup: KeyValueAny = !isLegacyEnabled(options) && ['QBKG03LM', 'QBKG22LM', 'QBKG04LM', 'QBKG21LM'].includes(model.model) ?
+                {0: 'hold', 1: 'release', 2: 'double'} : {0: 'single', 1: 'single'};
+
+            const action = actionLookup[msg.data['onOff']];
+            const button = mapping && mapping[msg.endpoint.ID] ? `_${mapping[msg.endpoint.ID]}` : '';
+
+            if (action === 'release') {
+                const anotherAction = globalStore.getValue(msg.endpoint, 'hold', false) ? 'hold_release' : 'single';
+                publish({action: `${anotherAction}${button}`});
+            }
+            globalStore.putValue(msg.endpoint, 'hold', action === 'hold');
+
+            return {action: `${action}${button}`};
+        },
+    } satisfies Fz.Converter,
+    lumi_occupancy_illuminance: {
+        // This is for occupancy sensor that only send a message when motion detected,
+        // but do not send a motion stop.
+        // Therefore we need to publish the no_motion detected by ourselves.
+        cluster: 'manuSpecificLumi',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.occupancy_timeout_2(), exposes.options.no_occupancy_since_true()],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data.hasOwnProperty('illuminance')) {
+                // The occupancy sensor only sends a message when motion detected.
+                // Therefore we need to publish the no_motion detected by ourselves.
+                let timeout = meta && meta.state && meta.state.hasOwnProperty('detection_interval') ?
+                    Number(meta.state.detection_interval) : ['RTCGQ14LM'].includes(model.model) ? 30 : 60;
+                timeout = options && options.hasOwnProperty('occupancy_timeout') && Number(options.occupancy_timeout) >= timeout ?
+                    Number(options.occupancy_timeout) : timeout + 2;
+
+                // Stop existing timers because motion is detected and set a new one.
+                clearTimeout(globalStore.getValue(msg.endpoint, 'occupancy_timer', null));
+
+                if (timeout !== 0) {
+                    const timer = setTimeout(() => {
+                        publish({occupancy: false});
+                    }, timeout * 1000);
+
+                    globalStore.putValue(msg.endpoint, 'occupancy_timer', timer);
+                }
+
+                // Sometimes RTCGQ14LM reports high illuminance values in the dark
+                // https://github.com/Koenkk/zigbee2mqtt/issues/12596
+                const illuminance = msg.data['illuminance'] > 130536 ? 0 : msg.data['illuminance'] - 65536;
+
+                const payload = {occupancy: true, illuminance};
+                noOccupancySince(msg.endpoint, options, publish, 'start');
+                return payload;
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_temperature: {
+        cluster: 'msTemperatureMeasurement',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const temperature = parseFloat(msg.data['measuredValue']) / 100.0;
+
+            // https://github.com/Koenkk/zigbee2mqtt/issues/798
+            // Sometimes the sensor publishes non-realistic vales.
+            if (temperature > -65 && temperature < 65) {
+                return {temperature};
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_curtain_position: {
+        cluster: 'genAnalogOutput',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.invert_cover()],
+        convert: (model, msg, publish, options, meta) => {
+            if ((model.model === 'ZNCLDJ12LM' || model.model === 'ZNCLDJ14LM') &&
+              msg.type === 'attributeReport' && [0, 2].includes(msg.data['presentValue'])) {
+                // Incorrect reports from the device, ignore (re-read by onEvent of ZNCLDJ12LM and ZNCLDJ14LM)
+                // https://github.com/Koenkk/zigbee-herdsman-converters/pull/1427#issuecomment-663862724
+                return;
+            }
+
+            let position = precisionRound(msg.data['presentValue'], 2);
+            position = options.invert_cover ? 100 - position : position;
+            return {position};
+        },
+    } satisfies Fz.Converter,
+    lumi_curtain_position_tilt: {
+        cluster: 'closuresWindowCovering',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.invert_cover()],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+            const invert = model.meta && model.meta.coverInverted ? !options.invert_cover : options.invert_cover;
+            if (msg.data.hasOwnProperty('currentPositionLiftPercentage') && msg.data['currentPositionLiftPercentage'] <= 100) {
+                const value = msg.data['currentPositionLiftPercentage'];
+                const position = invert ? 100 - value : value;
+                const state = invert ? (position > 0 ? 'CLOSE' : 'OPEN') : (position > 0 ? 'OPEN' : 'CLOSE');
+                result[postfixWithEndpointName('position', msg, model, meta)] = position;
+                result[postfixWithEndpointName('state', msg, model, meta)] = state;
+            }
+            if (msg.data.hasOwnProperty('currentPositionTiltPercentage') && msg.data['currentPositionTiltPercentage'] <= 100) {
+                const value = msg.data['currentPositionTiltPercentage'];
+                result[postfixWithEndpointName('tilt', msg, model, meta)] = invert ? 100 - value : value;
+            }
+            return result;
+        },
+    } satisfies Fz.Converter,
+    lumi_operation_mode_basic: {
+        cluster: 'genBasic',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const payload: KeyValueAny = {};
+
+            if (model.meta && !model.meta.multiEndpoint) {
+                const mappingMode: KeyValueNumberString = {0x12: 'control_relay', 0xFE: 'decoupled'};
+                const key = 0xFF22;
+                if (msg.data.hasOwnProperty(key)) {
+                    payload.operation_mode = mappingMode[msg.data[key]];
+                }
+            } else {
+                const mappingButton: KeyValueNumberString = {0xFF22: 'left', 0xFF23: 'right'};
+                const mappingMode: KeyValueNumberString = {0x12: 'control_left_relay', 0x22: 'control_right_relay', 0xFE: 'decoupled'};
+                for (const key in mappingButton) {
+                    if (msg.data.hasOwnProperty(key)) {
+                        payload[`operation_mode_${mappingButton[key]}`] = mappingMode[msg.data[key]];
+                    }
+                }
+            }
+
+            return payload;
+        },
+    } satisfies Fz.Converter,
+    lumi_bulb_interval: {
+        cluster: 'genBasic',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data['65281']) {
+                const data = msg.data['65281'];
+                return {
+                    state: data['100'] === 1 ? 'ON' : 'OFF',
+                    brightness: data['101'],
+                    color_temp: data['102'],
+                };
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_on_off_ignore_endpoint_4_5_6: {
+        cluster: 'genOnOff',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            // Lumi wall switches use endpoint 4, 5 or 6 to indicate an action on the button so we have to skip that.
+            if (msg.data.hasOwnProperty('onOff') && ![4, 5, 6].includes(msg.endpoint.ID)) {
+                const property = postfixWithEndpointName('state', msg, model, meta);
+                return {[property]: msg.data['onOff'] === 1 ? 'ON' : 'OFF'};
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_knob_rotation: {
+        cluster: 'manuSpecificLumi',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data.hasOwnProperty(570)) {
+                const act: KeyValueNumberString = {1: 'start_rotating', 2: 'rotation', 3: 'stop_rotating'};
+                return {
+                    action: act[msg.data[570]],
+                    action_rotation_angle: msg.data[558],
+                    action_rotation_angle_speed: msg.data[560],
+                    action_rotation_percent: msg.data[563],
+                    action_rotation_percent_speed: msg.data[562],
+                    action_rotation_time: msg.data[561],
+                };
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_curtain_status: {
+        cluster: 'genMultistateOutput',
+        type: ['attributeReport'],
+        convert: (model, msg, publish, options, meta) => {
+            let running = false;
+            const data = msg.data;
+            let lookup: KeyValueAny = {};
+
+            // For lumi.curtain.hagl04 and lumi.curtain.hagl07
+            if (['ZNCLDJ12LM', 'ZNCLDJ14LM'].includes(model.model)) lookup = {0: 'closing', 1: 'opening', 2: 'stop'};
+            // for lumi.curtain.acn002
+            if (['ZNJLBL01LM'].includes(model.model)) lookup = {0: 'declining', 1: 'rising', 2: 'pause', 3: 'blocked'};
+
+            if (data && data.hasOwnProperty('presentValue')) {
+                const value = data['presentValue'];
+                if (value < 2) {
+                    running = true;
+                }
+                return {
+                    motor_state: lookup[value],
+                    running: running,
+                };
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_WXKG01LM_action: {
+        // Unique converter
+        cluster: 'genOnOff',
+        type: ['attributeReport', 'readResponse'],
+        options: [
+            e.numeric('hold_timeout', ea.SET).withValueMin(0).withDescription(`The WXKG01LM only reports a button press and release.` +
+                `By default, a hold action is published when there is at least 1000 ms between both events. It could be that due to ` +
+                `delays in the network the release message is received late. This causes a single click to be identified as a hold ` +
+                `action. If you are experiencing this you can try experimenting with this option (e.g. set it to 2000) (value is in ms).`),
+            e.numeric('hold_timeout_expire', ea.SET).withValueMin(0).withDescription(`Sometimes it happens that the button does not send a ` +
+                `release. To avoid problems a release is automatically send after a timeout. The default timeout is 4000 ms, you can ` +
+                `increase it with this option (value is in ms).`),
+        ],
+        convert: (model, msg, publish, options: KeyValueAny, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            const state = msg.data['onOff'];
+
+            // 0 = click down, 1 = click up, else = multiple clicks
+            if (state === 0) {
+                const timer = setTimeout(() => {
+                    publish({action: 'hold'});
+                    globalStore.putValue(msg.endpoint, 'timer', null);
+                    globalStore.putValue(msg.endpoint, 'hold', Date.now());
+                    const holdTimer = setTimeout(() => {
+                        globalStore.putValue(msg.endpoint, 'hold', false);
+                    }, options.hold_timeout_expire || 4000);
+                    globalStore.putValue(msg.endpoint, 'hold_timer', holdTimer);
+                    // After 4000 milliseconds of not receiving release we assume it will not happen.
+                }, options.hold_timeout || 1000); // After 1000 milliseconds of not releasing we assume hold.
+                globalStore.putValue(msg.endpoint, 'timer', timer);
+            } else if (state === 1) {
+                if (globalStore.getValue(msg.endpoint, 'hold')) {
+                    const duration = Date.now() - globalStore.getValue(msg.endpoint, 'hold');
+                    publish({action: 'release', duration: duration});
+                    globalStore.putValue(msg.endpoint, 'hold', false);
+                }
+
+                if (globalStore.getValue(msg.endpoint, 'timer')) {
+                    clearTimeout(globalStore.getValue(msg.endpoint, 'timer'));
+                    globalStore.putValue(msg.endpoint, 'timer', null);
+                    publish({action: 'single'});
+                }
+            } else {
+                const clicks = msg.data['32768'];
+                const actionLookup: KeyValueAny = {1: 'single', 2: 'double', 3: 'triple', 4: 'quadruple'};
+                const payload = actionLookup[clicks] ? actionLookup[clicks] : 'many';
+                publish({action: payload});
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_WXKG11LM_action: {
+        cluster: 'genOnOff',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            let clicks;
+            if (msg.data.onOff) {
+                clicks = 1;
+            } else if (msg.data['32768']) {
+                clicks = msg.data['32768'];
+            }
+
+            const actionLookup: KeyValueAny = {1: 'single', 2: 'double', 3: 'triple', 4: 'quadruple'};
+            if (actionLookup[clicks]) {
+                return {action: actionLookup[clicks]};
+            }
+        },
+    } satisfies Fz.Converter,
+    RTCGQ13LM_occupancy: {
+        // This is for occupancy sensor that only send a message when motion detected,
+        // but do not send a motion stop.
+        // Therefore we need to publish the no_motion detected by ourselves.
+        cluster: 'msOccupancySensing',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.occupancy_timeout_2(), exposes.options.no_occupancy_since_true()],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data.occupancy !== 1) {
+                // In case of 0 no occupancy is reported.
+                // https://github.com/Koenkk/zigbee2mqtt/issues/467
+                return;
+            }
+
+            // The occupancy sensor only sends a message when motion detected.
+            // Therefore we need to publish the no_motion detected by ourselves.
+            let timeout: number = meta && meta.state && meta.state.hasOwnProperty('detection_interval') ?
+                Number(meta.state.detection_interval) : 60;
+            timeout = options && options.hasOwnProperty('occupancy_timeout') && Number(options.occupancy_timeout) >= timeout ?
+                Number(options.occupancy_timeout) : timeout + 2;
+
+            // Stop existing timers because motion is detected and set a new one.
+            clearTimeout(globalStore.getValue(msg.endpoint, 'occupancy_timer', null));
+
+            if (timeout !== 0) {
+                const timer = setTimeout(() => {
+                    publish({occupancy: false});
+                }, timeout * 1000);
+
+                globalStore.putValue(msg.endpoint, 'occupancy_timer', timer);
+            }
+
+            const payload = {occupancy: true};
+            noOccupancySince(msg.endpoint, options, publish, 'start');
+            return payload;
+        },
+    } satisfies Fz.Converter,
+    JTYJGD01LMBW_smoke: {
+        cluster: 'ssIasZone',
+        type: 'commandStatusChangeNotification',
+        convert: (model, msg, publish, options, meta) => {
+            const result = fz.ias_smoke_alarm_1.convert(model, msg, publish, options, meta);
+            const zoneStatus = msg.data.zonestatus;
+            if (result) result.test = (zoneStatus & 1<<1) > 0;
+            return result;
+        },
+    } satisfies Fz.Converter,
+    JTQJBF01LMBW_gas_density: {
+        cluster: 'genBasic',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const data = msg.data;
+            if (data && data['65281']) {
+                const basicAttrs = data['65281'];
+                if (basicAttrs.hasOwnProperty('100')) {
+                    return {gas_density: basicAttrs['100']};
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    JTQJBF01LMBW_sensitivity: {
+        cluster: 'ssIasZone',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const data = msg.data;
+            const lookup: KeyValueAny = {'1': 'low', '2': 'medium', '3': 'high'};
+
+            if (data && data.hasOwnProperty('65520')) {
+                const value = data['65520'];
+                if (value && value.startsWith('0x020')) {
+                    return {
+                        sensitivity: lookup[value.charAt(5)],
+                    };
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    DJT11LM_vibration: {
+        cluster: 'closuresDoorLock',
+        type: ['attributeReport', 'readResponse'],
+        options: [
+            exposes.options.vibration_timeout(),
+            exposes.options.calibration('x'),
+            exposes.options.calibration('y'),
+            exposes.options.calibration('z'),
+        ],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+
+            if (msg.data['85']) {
+                const vibrationLookup: KeyValueAny = {1: 'vibration', 2: 'tilt', 3: 'drop'};
+                result.action = vibrationLookup[msg.data['85']];
+
+                // Device only sends a message when vibration is detected.
+                // Therefore we need to publish a no_vibration message on our own.
+                if (result.action === 'vibration') {
+                    result.vibration = true;
+
+                    const timeout = options && options.hasOwnProperty('vibration_timeout') ? Number(options.vibration_timeout) : 90;
+
+                    // Stop any existing timer cause vibration detected
+                    clearTimeout(globalStore.getValue(msg.endpoint, 'vibration_timer', null));
+                    globalStore.putValue(msg.endpoint, 'vibration_timer', null);
+
+                    // Set new timer to publish no_vibration message
+                    if (timeout !== 0) {
+                        const timer = setTimeout(() => {
+                            publish({vibration: false});
+                        }, timeout * 1000);
+
+                        globalStore.putValue(msg.endpoint, 'vibration_timer', timer);
+                    }
+                }
+            }
+
+            if (msg.data['1283']) {
+                result.angle = msg.data['1283'];
+            }
+
+            if (msg.data['1285']) {
+                // https://github.com/dresden-elektronik/deconz-rest-plugin/issues/748#issuecomment-419669995
+                // Only first 2 bytes are relevant.
+                const data = (msg.data['1285'] >> 8);
+                // Swap byte order
+                result.strength = ((data & 0xFF) << 8) | ((data >> 8) & 0xFF);
+            }
+
+            if (msg.data['1288']) {
+                const data = msg.data['1288'];
+
+                // array interpretation:
+                // 12 bit two's complement sign extended integer
+                // data[1][bit0..bit15] : x
+                // data[1][bit16..bit31]: y
+                // data[0][bit0..bit15] : z
+                // left shift first to preserve sign extension for 'x'
+                let x = ((data['1'] << 16) >> 16);
+                let y = (data['1'] >> 16);
+                // left shift first to preserve sign extension for 'z'
+                let z = ((data['0'] << 16) >> 16);
+
+                // simple offset calibration
+                x=calibrateAndPrecisionRoundOptions(x, options, 'x');
+                y=calibrateAndPrecisionRoundOptions(y, options, 'y');
+                z=calibrateAndPrecisionRoundOptions(z, options, 'z');
+
+                // calibrated accelerometer values
+                result.x_axis=x;
+                result.y_axis=y;
+                result.z_axis=z;
+
+                // calculate angle
+                result.angle_x = Math.round(Math.atan(x/Math.sqrt(y*y+z*z)) * 180 / Math.PI);
+                result.angle_y = Math.round(Math.atan(y/Math.sqrt(x*x+z*z)) * 180 / Math.PI);
+                result.angle_z = Math.round(Math.atan(z/Math.sqrt(x*x+y*y)) * 180 / Math.PI);
+
+                // calculate absolute angle
+                const R = Math.sqrt(x * x + y * y + z * z);
+                result.angle_x_absolute = Math.round((Math.acos(x / R)) * 180 / Math.PI);
+                result.angle_y_absolute = Math.round((Math.acos(y / R)) * 180 / Math.PI);
+            }
+
+            return result;
+        },
+    } satisfies Fz.Converter,
+    DJT12LM_vibration: {
+        cluster: 'genOnOff',
+        type: 'commandOn',
+        convert: (model, msg, publish, options, meta) => {
+            return {action: 'vibration'};
+        },
+    } satisfies Fz.Converter,
+    ZNMS11LM_closuresDoorLock_report: {
+        cluster: 'closuresDoorLock',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+            const lockStatusLookup: KeyValueAny = {
+                1: 'finger_not_match',
+                2: 'password_not_match',
+                3: 'reverse_lock', // disable open from outside
+                4: 'reverse_lock_cancel', // enable open from outside
+                5: 'locked',
+                6: 'lock_opened',
+                7: 'finger_add',
+                8: 'finger_delete',
+                9: 'password_add',
+                10: 'password_delete',
+                11: 'lock_opened_inside', // Open form inside reverse lock enable
+                12: 'lock_opened_outside', // Open form outside reverse lock disable
+                13: 'ring_bell',
+                14: 'change_language_to',
+                15: 'finger_open',
+                16: 'password_open',
+                17: 'door_closed',
+            };
+            if (msg.data['65296']) { // finger/password success
+                const data = msg.data['65296'].toString(16);
+                const command = data.substr(0, 1); // 1 finger open, 2 password open
+                const userId = data.substr(5, 2);
+                const userType = data.substr(1, 1); // 1 admin, 2 user
+                result.data = data;
+                result.action = (lockStatusLookup[14+parseInt(command, 16)] +
+                    (userType === '1' ? '_admin' : '_user') + '_id' + parseInt(userId, 16).toString());
+                result.action_user = parseInt(userId, 16);
+            } else if (msg.data['65297']) { // finger, password failed or bell
+                const data = msg.data['65297'].toString(16);
+                const times = data.substr(0, 1);
+                const type = data.substr(5, 2); // 00 bell, 02 password, 40 error finger
+                result.data = data;
+                if (type === '40') {
+                    result.action_action = lockStatusLookup[1];
+                    result.action_repeat = parseInt(times, 16);
+                } else if (type === '02') {
+                    result.action = lockStatusLookup[2];
+                    result.action_repeat = parseInt(times, 16);
+                } else if (type === '00') {
+                    result.action = lockStatusLookup[13];
+                }
+            } else if (msg.data['65281'] && msg.data['65281']['1']) { // user added/delete
+                const data = msg.data['65281']['1'].toString(16);
+                const command = data.substr(0, 1); // 1 add, 2 delete
+                const userId = data.substr(5, 2);
+                result.data = data;
+                result.action = lockStatusLookup[6+parseInt(command, 16)];
+                result.action_user = parseInt(userId, 16);
+            }
+
+            if (isLegacyEnabled(options)) {
+                result.repeat = result.action_repeat;
+                result.user = result.action_user;
+            } else {
+                delete result.data;
+            }
+
+            return result;
+        },
+    } satisfies Fz.Converter,
+    ZNMS12LM_ZNMS13LM_closuresDoorLock_report: {
+        cluster: 'closuresDoorLock',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+            const lockStatusLookup: KeyValueAny = {
+                1: 'finger_not_match',
+                2: 'password_not_match',
+                3: 'reverse_lock', // disable open from outside
+                4: 'reverse_lock_cancel', // enable open from outside
+                5: 'locked',
+                6: 'lock_opened',
+                7: 'finger_add',
+                8: 'finger_delete',
+                9: 'password_add',
+                10: 'password_delete',
+                11: 'lock_opened_inside', // Open form inside reverse lock enable
+                12: 'lock_opened_outside', // Open form outside reverse lock disable
+                13: 'ring_bell',
+                14: 'change_language_to',
+                15: 'finger_open',
+                16: 'password_open',
+                17: 'door_closed',
+            };
+
+            if (msg.data['65526']) { // lock final status
+                // Convert data back to hex to decode
+                const data = Buffer.from(msg.data['65526'], 'ascii').toString('hex');
+                const command = data.substr(6, 4);
+                if (
+                    command === '0301' || // ZNMS12LM
+                        command === '0341' // ZNMS13LM
+                ) {
+                    result.action = lockStatusLookup[4];
+                    result.state = 'UNLOCK';
+                    result.reverse = 'UNLOCK';
+                } else if (
+                    command === '0311' || // ZNMS12LM
+                        command === '0351' // ZNMS13LM
+                ) {
+                    result.action = lockStatusLookup[4];
+                    result.state = 'LOCK';
+                    result.reverse = 'UNLOCK';
+                } else if (
+                    command === '0205' || // ZNMS12LM
+                        command === '0245' // ZNMS13LM
+                ) {
+                    result.action = lockStatusLookup[3];
+                    result.state = 'UNLOCK';
+                    result.reverse = 'LOCK';
+                } else if (
+                    command === '0215' || // ZNMS12LM
+                        command === '0255' || // ZNMS13LM
+                        command === '1355' // ZNMS13LM
+                ) {
+                    result.action = lockStatusLookup[3];
+                    result.state = 'LOCK';
+                    result.reverse = 'LOCK';
+                } else if (
+                    command === '0111' || // ZNMS12LM
+                        command === '1351' || // ZNMS13LM locked from inside
+                        command === '1451' // ZNMS13LM locked from outside
+                ) {
+                    result.action = lockStatusLookup[5];
+                    result.state = 'LOCK';
+                    result.reverse = 'UNLOCK';
+                } else if (
+                    command === '0b00' || // ZNMS12LM
+                        command === '0640' || // ZNMS13LM
+                        command === '0600' // ZNMS13LM
+
+                ) {
+                    result.action = lockStatusLookup[12];
+                    result.state = 'UNLOCK';
+                    result.reverse = 'UNLOCK';
+                } else if (
+                    command === '0c00' || // ZNMS12LM
+                        command === '2300' || // ZNMS13LM
+                        command === '0540' || // ZNMS13LM
+                        command === '0440' // ZNMS13LM
+                ) {
+                    result.action = lockStatusLookup[11];
+                    result.state = 'UNLOCK';
+                    result.reverse = 'UNLOCK';
+                } else if (
+                    command === '2400' || // ZNMS13LM door closed from insed
+                        command === '2401' // ZNMS13LM door closed from outside
+                ) {
+                    result.action = lockStatusLookup[17];
+                    result.state = 'UNLOCK';
+                    result.reverse = 'UNLOCK';
+                }
+            } else if (msg.data['65296']) { // finger/password success
+                const data = Buffer.from(msg.data['65296'], 'ascii').toString('hex');
+                const command = data.substr(6, 2); // 1 finger open, 2 password open
+                const userId = data.substr(12, 2);
+                const userType = data.substr(8, 1); // 1 admin, 2 user
+                result.action = (lockStatusLookup[14+parseInt(command, 16)] +
+                    (userType === '1' ? '_admin' : '_user') + '_id' + parseInt(userId, 16).toString());
+                result.action_user = parseInt(userId, 16);
+            } else if (msg.data['65297']) { // finger, password failed or bell
+                const data = Buffer.from(msg.data['65297'], 'ascii').toString('hex');
+                const times = data.substr(6, 2);
+                const type = data.substr(12, 2); // 00 bell, 02 password, 40 error finger
+                if (type === '40') {
+                    result.action = lockStatusLookup[1];
+                    result.action_repeat = parseInt(times, 16);
+                } else if (type === '00') {
+                    result.action = lockStatusLookup[13];
+                    result.action_repeat = null;
+                } else if (type === '02') {
+                    result.action = lockStatusLookup[2];
+                    result.action_repeat = parseInt(times, 16);
+                }
+            } else if (msg.data['65281']) { // password added/delete
+                const data = Buffer.from(msg.data['65281'], 'ascii').toString('hex');
+                const command = data.substr(18, 2); // 1 add, 2 delete
+                const userId = data.substr(12, 2);
+                result.action = lockStatusLookup[6+parseInt(command, 16)];
+                result.action_user = parseInt(userId, 16);
+            } else if (msg.data['65522']) { // set language
+                const data = Buffer.from(msg.data['65522'], 'ascii').toString('hex');
+                const langId = data.substr(6, 2); // 1 chinese, 2: english
+                result.action = (lockStatusLookup[14])+ (langId==='2'?'_english':'_chinese');
+            }
+
+            if (isLegacyEnabled(options)) {
+                result.repeat = result.action_repeat;
+                result.user = result.action_user;
+            }
+
+            return result;
+        },
+    } satisfies Fz.Converter,
+    ZNMS12LM_low_battery: {
+        cluster: 'genPowerCfg',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data.hasOwnProperty('batteryAlarmMask')) {
+                return {battery_low: msg.data['batteryAlarmMask'] === 1};
+            }
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            const actionLookup: KeyValueAny = {0: 'hold', 255: 'release', 1: 'single', 2: 'double', 3: 'triple', 5: 'quintuple', 6: 'many'};
+            const button = msg.endpoint.ID;
+            const value = msg.data.presentValue;
+            clearTimeout(globalStore.getValue(msg.endpoint, 'timer'));
+            if (model.model === 'WXKG13LM') {
+                return {action: `${actionLookup[value]}`};
+            } else {
+                // 0 = hold
+                if (value === 0) {
+                    // Aqara Opple does not generate a release event when pressed for more than 5 seconds
+                    // After 5 seconds of not releasing we assume release.
+                    const timer = setTimeout(() => publish({action: `button_${button}_release`}), 5000);
+                    globalStore.putValue(msg.endpoint, 'timer', timer);
+                }
+                return {action: `button_${button}_${actionLookup[value]}`};
+            }
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_on: {
+        cluster: 'genOnOff',
+        type: 'commandOn',
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            return {action: 'button_2_single'};
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_off: {
+        cluster: 'genOnOff',
+        type: 'commandOff',
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            return {action: 'button_1_single'};
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_step: {
+        cluster: 'genLevelCtrl',
+        type: 'commandStep',
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            const button = msg.data.stepmode === 0 ? '4' : '3';
+            return {action: `button_${button}_single`};
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_stop: {
+        cluster: 'genLevelCtrl',
+        type: 'commandStop',
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            if (globalStore.hasValue(msg.endpoint, 'button')) {
+                const value = globalStore.getValue(msg.endpoint, 'button');
+                const duration = Date.now() - value.start;
+                const payload = {action: `button_${value.button}_release`, duration, action_duration: duration};
+                if (!isLegacyEnabled(options)) delete payload.duration;
+                return payload;
+            }
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_move: {
+        cluster: 'genLevelCtrl',
+        type: 'commandMove',
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            const button = msg.data.movemode === 0 ? '4' : '3';
+            globalStore.putValue(msg.endpoint, 'button', {button, start: Date.now()});
+            return {action: `button_${button}_hold`};
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_step_color_temp: {
+        cluster: 'lightingColorCtrl',
+        type: 'commandStepColorTemp',
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            let action;
+            if (model.model === 'WXCJKG12LM') {
+                // for WXCJKG12LM model it's double click event on buttons 3 and 4
+                action = (msg.data.stepmode === 1) ? '3_double' : '4_double';
+            } else {
+                // but for WXCJKG13LM model it's single click event on buttons 5 and 6
+                action = (msg.data.stepmode === 1) ? '5_single' : '6_single';
+            }
+            return {action: `button_${action}`};
+        },
+    } satisfies Fz.Converter,
+    aqara_opple_move_color_temp: {
+        cluster: 'lightingColorCtrl',
+        type: 'commandMoveColorTemp',
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (hasAlreadyProcessedMessage(msg, model)) return;
+            const stop = msg.data.movemode === 0;
+            let result = null;
+            if (stop) {
+                const button = globalStore.getValue(msg.endpoint, 'button').button;
+                const duration = Date.now() - globalStore.getValue(msg.endpoint, 'button').start;
+                result = {action: `button_${button}_release`, duration, action_duration: duration};
+                if (!isLegacyEnabled(options)) delete result.duration;
+            } else {
+                const button = msg.data.movemode === 3 ? '6' : '5';
+                result = {action: `button_${button}_hold`};
+                globalStore.putValue(msg.endpoint, 'button', {button, start: Date.now()});
+            }
+            return result;
+        },
+    } satisfies Fz.Converter,
+    ZNCJMB14LM: {
+        cluster: 'manuSpecificLumi',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+            if (msg.data.hasOwnProperty(0x0215)) {
+                const lookup: KeyValueAny = {0: 'classic', 1: 'concise'};
+                result.theme = lookup[msg.data[0x0215]];
+            }
+            if (msg.data.hasOwnProperty(0x0214)) {
+                const lookup: KeyValueAny = {1: 'classic', 2: 'analog clock'};
+                result.screen_saver_style = lookup[msg.data[0x0214]];
+            }
+            if (msg.data.hasOwnProperty(0x0213)) {
+                result.standby_enabled = msg.data[0x0213] & 1 ? true : false;
+            }
+            if (msg.data.hasOwnProperty(0x0212)) {
+                const lookup: KeyValueAny = {0: 'mute', 1: 'low', 2: 'medium', 3: 'high'};
+                result.beep_volume = lookup[msg.data[0x0212]];
+            }
+            if (msg.data.hasOwnProperty(0x0211)) {
+                result.lcd_brightness = msg.data[0x0211];
+            }
+            if (msg.data.hasOwnProperty(0x022b)) {
+                const lookup: KeyValueAny = {0: 'none', 1: '1', 2: '2', 3: '1 and 2', 4: '3', 5: '1 and 3', 6: '2 and 3', 7: 'all'};
+                result.available_switches = lookup[msg.data[0x022b]];
+            }
+            if (msg.data.hasOwnProperty(0x217)) {
+                const lookup: KeyValueAny = {3: 'small', 4: 'medium', 5: 'large'};
+                result.font_size = lookup[msg.data[0x217]];
+            }
+            if (msg.data.hasOwnProperty(0x219)) {
+                const lookup: KeyValueAny = {0: 'scene', 1: 'feel', 2: 'thermostat', 3: 'switch'};
+                result.homepage = lookup[msg.data[0x219]];
+            }
+            if (msg.data.hasOwnProperty(0x210)) {
+                const lookup: KeyValueAny = {0: 'chinese', 1: 'english'};
+                result.language = lookup[msg.data[0x210]];
+            }
+            if (msg.data.hasOwnProperty(0x216)) {
+                result.standby_time = msg.data[0x216];
+            }
+            if (msg.data.hasOwnProperty(0x218)) {
+                result.lcd_auto_brightness_enabled = msg.data[0x218] & 1 ? true : false;
+            }
+            if (msg.data.hasOwnProperty(0x221)) {
+                result.screen_saver_enabled = msg.data[0x221] & 1 ? true : false;
+            }
+            if (msg.data.hasOwnProperty(0x222)) {
+                result.standby_lcd_brightness = msg.data[0x222];
+            }
+            if (msg.data.hasOwnProperty(0x223)) {
+                const lookup: KeyValueAny = {1: '1', 2: '2', 3: '3', 4: '4', 5: '5', 6: '6', 7: '7', 8: '8', 9: '9', 10: '10', 11: '11'};
+                const textarr = msg.data[0x223].slice(1, msg.data[0x223].length);
+                result.switch_1_icon = lookup[msg.data[0x223][0]];
+                result.switch_1_text = String.fromCharCode(...textarr);
+            }
+            if (msg.data.hasOwnProperty(0x224)) {
+                const lookup: KeyValueAny = {1: '1', 2: '2', 3: '3', 4: '4', 5: '5', 6: '6', 7: '7', 8: '8', 9: '9', 10: '10', 11: '11'};
+                const textarr = msg.data[0x224].slice(1, msg.data[0x224].length);
+                result.switch_2_icon = lookup[msg.data[0x224][0]];
+                result.switch_2_text = String.fromCharCode(...textarr);
+            }
+            if (msg.data.hasOwnProperty(0x225)) {
+                const lookup: KeyValueAny = {1: '1', 2: '2', 3: '3', 4: '4', 5: '5', 6: '6', 7: '7', 8: '8', 9: '9', 10: '10', 11: '11'};
+                const textarr = msg.data[0x225].slice(1, msg.data[0x225].length);
+                result.switch_3_icon = lookup[msg.data[0x225][0]];
+                result.switch_3_text = String.fromCharCode(...textarr);
+            }
+            return result;
+        },
+    } satisfies Fz.Converter,
+    lumi_cube_action_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            /*
+            Source: https://github.com/kirovilya/ioBroker.zigbee
+                +---+
+                | 2 |
+            +---+---+---+
+            | 4 | 0 | 1 |
+            +---+---+---+
+                |M5I|
+                +---+
+                | 3 |
+                +---+
+            Side 5 is with the MI logo, side 3 contains the battery door.
+            presentValue = 0 = shake
+            presentValue = 2 = wakeup
+            presentValue = 3 = fly/fall
+            presentValue = y + x * 8 + 64 = 90º Flip from side x on top to side y on top
+            presentValue = x + 128 = 180º flip to side x on top
+            presentValue = x + 256 = push/slide cube while side x is on top
+            presentValue = x + 512 = double tap while side x is on top
+            */
+            const value = msg.data['presentValue'];
+            let result = null;
+
+            if (value === 0) result = {action: 'shake'};
+            else if (value === 1) result = {action: 'throw'};
+            else if (value === 2) result = {action: 'wakeup'};
+            else if (value === 3) result = {action: 'fall'};
+            else if (value >= 512) result = {action: 'tap', side: value-512};
+            else if (value >= 256) result = {action: 'slide', side: value-256};
+            else if (value >= 128) result = {action: 'flip180', side: value-128};
+            else if (value >= 64) {
+                result = {
+                    action: 'flip90', action_from_side: Math.floor((value-64) / 8), action_to_side: value % 8, action_side: value % 8,
+                    from_side: Math.floor((value-64) / 8), to_side: value % 8, side: value % 8,
+                };
+            }
+
+            if (result && !isLegacyEnabled(options)) {
+                delete result.to_side;
+                delete result.from_side;
+            }
+
+            return result ? result : null;
+        },
+    } satisfies Fz.Converter,
+    lumi_cube_action_analog: {
+        cluster: 'genAnalogInput',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            /*
+            Source: https://github.com/kirovilya/ioBroker.zigbee
+            presentValue = rotation angle left < 0, right > 0
+            */
+            const value = msg.data['presentValue'];
+            const result: KeyValueAny = {
+                action: value < 0 ? 'rotate_left' : 'rotate_right',
+                angle: Math.floor(value * 100) / 100,
+                action_angle: Math.floor(value * 100) / 100,
+            };
+
+            if (!isLegacyEnabled(options)) delete result.angle;
+            return result;
+        },
+    } satisfies Fz.Converter,
+    RTCGQ11LM_illuminance: {
+        cluster: 'msIlluminanceMeasurement',
+        type: ['attributeReport', 'readResponse'],
+        convert: (model, msg, publish, options, meta) => {
+            // also trigger movement, because there is no illuminance without movement
+            // https://github.com/Koenkk/zigbee-herdsman-converters/issues/1925
+            msg.data.occupancy = 1;
+            const payload = fz.occupancy_with_timeout.convert(model, msg, publish, options, meta) as KeyValueAny;
+            if (payload) {
+                // DEPRECATED: remove illuminance_lux here.
+                const illuminance = msg.data['measuredValue'];
+                payload.illuminance = illuminance;
+                payload.illuminance_lux = illuminance;
+            }
+            return payload;
+        },
+    } satisfies Fz.Converter,
 };
 
 export const toZigbee = {
-    aqara_feeder: {
+    lumi_feeder: {
         key: ['feed', 'schedule', 'led_indicator', 'child_lock', 'mode', 'serving_size', 'portion_weight'],
         convertSet: async (entity, key, value, meta) => {
             const sendAttr = async (attrCode: number, value: number, length: number) => {
@@ -1670,6 +2895,1330 @@ export const toZigbee = {
             return {state: {[key]: value}};
         },
     } satisfies Tz.Converter,
+    lumi_detection_distance: {
+        key: ['detection_distance'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, 'detection_distance');
+            value = value.toLowerCase();
+            const lookup = {'10mm': 1, '20mm': 2, '30mm': 3};
+            await entity.write('manuSpecificLumi', {0x010C: {value: getFromLookup(value, lookup), type: 0x20}}, {manufacturerCode});
+            return {state: {detection_distance: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x010C], {manufacturerCode});
+        },
+    } satisfies Tz.Converter,
+    lumi_trv: {
+        key: ['system_mode', 'preset', 'window_detection', 'valve_detection', 'child_lock', 'away_preset_temperature',
+            'calibrate', 'sensor', 'external_temperature_input', 'identify', 'schedule', 'schedule_settings'],
+        convertSet: async (entity, key, value, meta) => {
+            const lumiHeader = (counter: number, params: number[], action: number) => {
+                const header = [0xaa, 0x71, params.length + 3, 0x44, counter];
+                const integrity = 512 - header.reduce((sum, elem) => sum + elem, 0);
+                return [...header, integrity, action, 0x41, params.length];
+            };
+            const sensor = Buffer.from('00158d00019d1b98', 'hex');
+
+            switch (key) {
+            case 'system_mode':
+                await entity.write('manuSpecificLumi', {0x0271: {value: getFromLookup(value, {'off': 0, 'heat': 1}), type: 0x20}},
+                    {manufacturerCode: manufacturerCode});
+                break;
+            case 'preset':
+                await entity.write('manuSpecificLumi', {0x0272: {value: getFromLookup(value, {'manual': 0, 'auto': 1, 'away': 2}), type: 0x20}},
+                    {manufacturerCode: manufacturerCode});
+                break;
+            case 'window_detection':
+                await entity.write('manuSpecificLumi', {
+                    0x0273: {value: getFromLookup(value, {'false': 0, 'true': 1}, undefined, true), type: 0x20},
+                }, {manufacturerCode: manufacturerCode});
+                break;
+            case 'valve_detection':
+                await entity.write('manuSpecificLumi', {
+                    0x0274: {value: getFromLookup(value, {'false': 0, 'true': 1}, undefined, true), type: 0x20},
+                }, {manufacturerCode: manufacturerCode});
+                break;
+            case 'child_lock':
+                await entity.write('manuSpecificLumi', {
+                    0x0277: {value: getFromLookup(value, {'false': 0, 'true': 1}, undefined, true), type: 0x20},
+                }, {manufacturerCode: manufacturerCode});
+                break;
+            case 'away_preset_temperature':
+                await entity.write('manuSpecificLumi', {
+                    0x0279: {value: Math.round(toNumber(value, 'away_preset_temperature') * 100), type: 0x23},
+                }, {manufacturerCode: manufacturerCode});
+                break;
+            case 'sensor': {
+                assertEndpoint(entity);
+                const device = Buffer.from(entity.deviceIeeeAddress.substring(2), 'hex');
+                const timestamp = Buffer.alloc(4);
+                timestamp.writeUint32BE(Date.now()/1000);
+
+                if (value === 'external') {
+                    const params1 = [
+                        ...timestamp,
+                        0x3d, 0x04,
+                        ...device,
+                        ...sensor,
+                        0x00, 0x01, 0x00, 0x55,
+                        0x13, 0x0a, 0x02, 0x00, 0x00, 0x64, 0x04, 0xce, 0xc2, 0xb6, 0xc8,
+                        0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x3d,
+                        0x64,
+                        0x65,
+                    ];
+                    const params2 = [
+                        ...timestamp,
+                        0x3d, 0x05,
+                        ...device,
+                        ...sensor,
+                        0x08, 0x00, 0x07, 0xfd,
+                        0x16, 0x0a, 0x02, 0x0a, 0xc9, 0xe8, 0xb1, 0xb8, 0xd4, 0xda, 0xcf, 0xdf, 0xc0, 0xeb,
+                        0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x3d,
+                        0x04,
+                        0x65,
+                    ];
+
+                    const val1 = [...(lumiHeader(0x12, params1, 0x02)), ...params1];
+                    const val2 = [...(lumiHeader(0x13, params2, 0x02)), ...params2];
+
+                    await entity.write('manuSpecificLumi', {0xfff2: {value: val1, type: 0x41}}, {manufacturerCode: manufacturerCode});
+                    await entity.write('manuSpecificLumi', {0xfff2: {value: val2, type: 0x41}}, {manufacturerCode: manufacturerCode});
+                } else if (value === 'internal') {
+                    const params1 = [
+                        ...timestamp,
+                        0x3d, 0x05,
+                        ...device,
+                        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                    ];
+                    const params2 = [
+                        ...timestamp,
+                        0x3d, 0x04,
+                        ...device,
+                        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                    ];
+
+                    const val1 = [...(lumiHeader(0x12, params1, 0x04)), ...params1];
+                    const val2 = [...(lumiHeader(0x13, params2, 0x04)), ...params2];
+
+                    await entity.write('manuSpecificLumi', {0xfff2: {value: val1, type: 0x41}}, {manufacturerCode: manufacturerCode});
+                    await entity.write('manuSpecificLumi', {0xfff2: {value: val2, type: 0x41}}, {manufacturerCode: manufacturerCode});
+
+                    await entity.read('hvacThermostat', ['localTemp']);
+                }
+                break;
+            }
+            case 'external_temperature_input':
+                if (meta.state['sensor'] === 'external') {
+                    const temperatureBuf = Buffer.alloc(4);
+                    const number = toNumber(value);
+                    temperatureBuf.writeFloatBE(Math.round(number * 100));
+
+                    const params = [...sensor, 0x00, 0x01, 0x00, 0x55, ...temperatureBuf];
+                    const data = [...(lumiHeader(0x12, params, 0x05)), ...params];
+
+                    await entity.write('manuSpecificLumi', {0xfff2: {value: data, type: 0x41}}, {manufacturerCode: manufacturerCode});
+                }
+                break;
+            case 'calibrate':
+                await entity.write('manuSpecificLumi', {0x0270: {value: 1, type: 0x20}}, {manufacturerCode: 0x115F});
+                break;
+            case 'identify':
+                await entity.command('genIdentify', 'identify', {identifytime: 5}, {});
+                break;
+            case 'schedule':
+                await entity.write('manuSpecificLumi', {
+                    0x027d: {value: getFromLookup(value, {'false': 0, 'true': 1}, undefined, true), type: 0x20},
+                }, {manufacturerCode: manufacturerCode});
+                break;
+            case 'schedule_settings': {
+                // @ts-expect-error
+                const schedule = trv.parseSchedule(value);
+                trv.validateSchedule(schedule);
+                const buffer = trv.encodeSchedule(schedule);
+                await entity.write('manuSpecificLumi', {0x0276: {value: buffer, type: 0x41}}, {manufacturerCode: manufacturerCode});
+                break;
+            }
+            default: // Unknown key
+                meta.logger.warn(`zigbee-herdsman-converters:lumi_trv: Unhandled key ${key}`);
+            }
+        },
+        convertGet: async (entity, key, meta) => {
+            const dict = {'system_mode': 0x0271, 'preset': 0x0272, 'window_detection': 0x0273, 'valve_detection': 0x0274,
+                'child_lock': 0x0277, 'away_preset_temperature': 0x0279, 'calibrated': 0x027b, 'sensor': 0x027e,
+                'schedule': 0x027d, 'schedule_settings': 0x0276};
+
+            if (dict.hasOwnProperty(key)) {
+                await entity.read('manuSpecificLumi', [getFromLookup(key, dict)], {manufacturerCode: 0x115F});
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_presence_region_upsert: {
+        key: ['region_upsert'],
+        convertSet: async (entity, key, value, meta) => {
+            const log = createLogger(meta.logger, 'lumi', 'lumi_presence:region_upsert');
+            const commandWrapper = presence.parseAqaraFp1RegionUpsertInput(value);
+
+            if (!commandWrapper.isSuccess) {
+                log('warn',
+                    // @ts-expect-error
+                    `encountered an error (${commandWrapper.error.reason}) ` +
+                    `while parsing configuration commands (input: ${JSON.stringify(value)})`,
+                );
+
+                return;
+            }
+
+            const command = commandWrapper.payload.command;
+
+            log('debug', `trying to create region ${command.region_id}`);
+
+            const sortedZonesAccumulator = {};
+            const sortedZonesWithSets: {[s: number]: [number]} = command.zones
+                .reduce(
+                    (accumulator: {[s: number]: Set<number>}, zone: {x: number, y: number}) => {
+                        if (!accumulator[zone.y]) {
+                            accumulator[zone.y] = new Set<number>();
+                        }
+
+                        accumulator[zone.y].add(zone.x);
+
+                        return accumulator;
+                    },
+                    sortedZonesAccumulator,
+                );
+            const sortedZones = Object.entries(sortedZonesWithSets).reduce((acc, [key, value]) => {
+                const numKey = parseInt(key, 10); // Convert string key back to number
+                acc[numKey] = Array.from(value);
+                return acc;
+            }, {} as {[s: number]: number[]});
+
+            const deviceConfig = new Uint8Array(7);
+
+            // Command parameters
+            deviceConfig[0] = presence.constants.region_config_cmds.create;
+            deviceConfig[1] = command.region_id;
+            deviceConfig[6] = presence.constants.region_config_cmd_suffix_upsert;
+            // Zones definition
+            deviceConfig[2] |= presence.encodeXCellsDefinition(sortedZones['1']);
+            deviceConfig[2] |= presence.encodeXCellsDefinition(sortedZones['2']) << 4;
+            deviceConfig[3] |= presence.encodeXCellsDefinition(sortedZones['3']);
+            deviceConfig[3] |= presence.encodeXCellsDefinition(sortedZones['4']) << 4;
+            deviceConfig[4] |= presence.encodeXCellsDefinition(sortedZones['5']);
+            deviceConfig[4] |= presence.encodeXCellsDefinition(sortedZones['6']) << 4;
+            deviceConfig[5] |= presence.encodeXCellsDefinition(sortedZones['7']);
+
+            log('info', `create region ${command.region_id} ${printNumbersAsHexSequence([...deviceConfig], 2)}`);
+
+            const payload = {
+                [presence.constants.region_config_write_attribute]: {
+                    value: deviceConfig,
+                    type: presence.constants.region_config_write_attribute_type,
+                },
+            };
+
+            await entity.write('manuSpecificLumi', payload, {manufacturerCode});
+        },
+    } satisfies Tz.Converter,
+    lumi_presence_region_delete: {
+        key: ['region_delete'],
+        convertSet: async (entity, key, value, meta) => {
+            const log = createLogger(meta.logger, 'lumi', 'lumi_presence:region_delete');
+            const commandWrapper = presence.parseAqaraFp1RegionDeleteInput(value);
+
+            if (!commandWrapper.isSuccess) {
+                log('warn',
+                    // @ts-expect-error
+                    `encountered an error (${commandWrapper.error.reason}) ` +
+                    `while parsing configuration commands (input: ${JSON.stringify(value)})`,
+                );
+                return;
+            }
+            const command = commandWrapper.payload.command;
+
+            log('debug', `trying to delete region ${command.region_id}`);
+
+            const deviceConfig = new Uint8Array(7);
+
+            // Command parameters
+            deviceConfig[0] = presence.constants.region_config_cmds.delete;
+            deviceConfig[1] = command.region_id;
+            deviceConfig[6] = presence.constants.region_config_cmd_suffix_delete;
+            // Zones definition
+            deviceConfig[2] = 0;
+            deviceConfig[3] = 0;
+            deviceConfig[4] = 0;
+            deviceConfig[5] = 0;
+
+            log('info',
+                `delete region ${command.region_id} ` +
+                `(${printNumbersAsHexSequence([...deviceConfig], 2)})`,
+            );
+
+            const payload = {
+                [presence.constants.region_config_write_attribute]: {
+                    value: deviceConfig,
+                    type: presence.constants.region_config_write_attribute_type,
+                },
+            };
+
+            await entity.write('manuSpecificLumi', payload, {manufacturerCode});
+        },
+    } satisfies Tz.Converter,
+    lumi_cube_t1_operation_mode: {
+        key: ['operation_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {action_mode: 0, scene_mode: 1};
+            /**
+             * schedule the callback to run when the configuration window comes
+             */
+            const callback = async () => {
+                await entity.write(
+                    'manuSpecificLumi',
+                    {0x0148: {value: getFromLookup(value, lookup), type: 0x20}},
+                    {manufacturerCode: manufacturerCode, disableDefaultResponse: true},
+                );
+                meta.logger.info('operation_mode switch success!');
+            };
+            globalStore.putValue(meta.device, 'opModeSwitchTask', {callback, newMode: value});
+            meta.logger.info('Now give your cube a forceful throw motion (Careful not to drop it)!');
+        },
+    } satisfies Tz.Converter,
+    lumi_detection_interval: {
+        key: ['detection_interval'],
+        convertSet: async (entity, key, value, meta) => {
+            assertNumber(value, key);
+            value *= 1;
+            await entity.write('manuSpecificLumi', {0x0102: {value: [value], type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {detection_interval: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0102], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_overload_protection: {
+        key: ['overload_protection'],
+        convertSet: async (entity, key, value, meta) => {
+            assertNumber(value, key);
+            value *= 1;
+            await entity.write('manuSpecificLumi', {0x020b: {value: [value], type: 0x39}}, manufacturerOptions.lumi);
+            return {state: {overload_protection: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x020b], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_mode_switch: {
+        key: ['mode_switch'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'anti_flicker_mode': 4, 'quick_mode': 1};
+            await entity.write('manuSpecificLumi', {0x0004: {value: getFromLookup(value, lookup), type: 0x21}}, manufacturerOptions.lumi);
+            return {state: {mode_switch: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0004], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_button_switch_mode: {
+        key: ['button_switch_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'relay': 0, 'relay_and_usb': 1};
+            await entity.write('manuSpecificLumi', {0x0226: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {button_switch_mode: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0226], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_socket_button_lock: {
+        key: ['button_lock'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'ON': 0, 'OFF': 1};
+            await entity.write('manuSpecificLumi', {0x0200: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {button_lock: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0200], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_led_disabled_night: {
+        key: ['led_disabled_night'],
+        convertSet: async (entity, key, value, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['ZNCZ04LM', 'ZNCZ12LM', 'ZNCZ15LM', 'QBCZ14LM', 'QBCZ15LM', 'QBKG19LM', 'QBKG18LM', 'QBKG20LM', 'QBKG25LM', 'QBKG26LM',
+                'QBKG28LM', 'QBKG29LM', 'QBKG30LM', 'QBKG31LM', 'QBKG32LM', 'QBKG34LM', 'DLKZMK11LM', 'SSM-U01', 'WS-EUK01', 'WS-EUK02',
+                'WS-EUK03', 'WS-EUK04', 'SP-EUC01', 'ZNQBKG24LM', 'ZNQBKG25LM'].includes(meta.mapped.model)) {
+                await entity.write('manuSpecificLumi', {0x0203: {value: value ? 1 : 0, type: 0x10}}, manufacturerOptions.lumi);
+            } else if (['ZNCZ11LM'].includes(meta.mapped.model)) {
+                const payload = value ?
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x00, 0x03, 0x10, 0x00] :
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x01, 0x03, 0x10, 0x01];
+
+                await entity.write('genBasic', {0xFFF0: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+            } else {
+                throw new Error('Not supported');
+            }
+            return {state: {led_disabled_night: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['ZNCZ04LM', 'ZNCZ12LM', 'ZNCZ15LM', 'QBCZ15LM', 'QBCZ14LM', 'QBKG19LM', 'QBKG18LM', 'QBKG20LM', 'QBKG25LM', 'QBKG26LM',
+                'QBKG28LM', 'QBKG29LM', 'QBKG30LM', 'QBKG31LM', 'QBKG32LM', 'QBKG34LM', 'DLKZMK11LM', 'SSM-U01', 'WS-EUK01', 'WS-EUK02',
+                'WS-EUK03', 'WS-EUK04', 'SP-EUC01', 'ZNQBKG24LM', 'ZNQBKG25LM'].includes(meta.mapped.model)) {
+                await entity.read('manuSpecificLumi', [0x0203], manufacturerOptions.lumi);
+            } else {
+                throw new Error('Not supported');
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_flip_indicator_light: {
+        key: ['flip_indicator_light'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'OFF': 0, 'ON': 1};
+            await entity.write('manuSpecificLumi', {0x00F0: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {flip_indicator_light: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x00F0], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_dimmer_mode: {
+        key: ['dimmer_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'rgbw': 3, 'dual_ct': 1};
+            assertString(value, key);
+            value = value.toLowerCase();
+            // @ts-expect-error
+            if (['rgbw'].includes(value)) {
+                await entity.write('manuSpecificLumi', {0x0509: {value: getFromLookup(value, lookup), type: 0x23}}, manufacturerOptions.lumi);
+                await entity.write('manuSpecificLumi', {0x050F: {value: 1, type: 0x23}}, manufacturerOptions.lumi);
+            } else {
+                await entity.write('manuSpecificLumi', {0x0509: {value: getFromLookup(value, lookup), type: 0x23}}, manufacturerOptions.lumi);
+                // Turn on dimming channel 1 and channel 2
+                await entity.write('manuSpecificLumi', {0x050F: {value: 3, type: 0x23}}, manufacturerOptions.lumi);
+            }
+            return {state: {dimmer_mode: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0509], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_operation_mode_basic: {
+        key: ['operation_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            assertEndpoint(entity);
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            let targetValue = isObject(value) && value.hasOwnProperty('state') ? value.state : value;
+
+            // 1/2 gang switches using genBasic on endpoint 1.
+            let attrId;
+            let attrValue: number;
+            if (meta.mapped.meta && meta.mapped.meta.multiEndpoint) {
+                attrId = {left: 0xFF22, right: 0xFF23}[meta.endpoint_name];
+                // Allow usage of control_relay for 2 gang switches by mapping it to the default side.
+                if (targetValue === 'control_relay') {
+                    targetValue = `control_${meta.endpoint_name}_relay`;
+                }
+                attrValue = getFromLookup(targetValue, {control_left_relay: 0x12, control_right_relay: 0x22, decoupled: 0xFE});
+
+                if (attrId == null) {
+                    throw new Error(`Unsupported endpoint ${meta.endpoint_name} for changing operation_mode.`);
+                }
+            } else {
+                attrId = 0xFF22;
+                attrValue = getFromLookup(targetValue, {control_relay: 0x12, decoupled: 0xFE});
+            }
+
+            if (attrValue == null) {
+                throw new Error('Invalid operation_mode value');
+            }
+
+            const endpoint = entity.getDevice().getEndpoint(1);
+            const payload: KeyValueAny = {};
+            payload[attrId] = {value: attrValue, type: 0x20};
+            await endpoint.write('genBasic', payload, manufacturerOptions.lumi);
+
+            return {state: {operation_mode: targetValue}};
+        },
+        convertGet: async (entity, key, meta) => {
+            let attrId;
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (meta.mapped.meta && meta.mapped.meta.multiEndpoint) {
+                attrId = {left: 0xFF22, right: 0xFF23}[meta.endpoint_name];
+                if (attrId == null) {
+                    throw new Error(`Unsupported endpoint ${meta.endpoint_name} for getting operation_mode.`);
+                }
+            } else {
+                attrId = 0xFF22;
+            }
+            await entity.read('genBasic', [attrId], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_operation_mode_opple: {
+        key: ['operation_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            // Support existing syntax of a nested object just for the state field. Though it's quite silly IMO.
+            const targetValue = isObject(value) && value.hasOwnProperty('state') ? value.state : value;
+            // Switches using manuSpecificLumi 0x0200 on the same endpoints as the onOff clusters.
+            const lookupState = {control_relay: 0x01, decoupled: 0x00};
+            await entity.write('manuSpecificLumi', {0x0200:
+                {value: getFromLookup(targetValue, lookupState), type: 0x20}}, manufacturerOptions.lumi);
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0200], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_do_not_disturb: {
+        key: ['do_not_disturb'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x0203: {value: value ? 1 : 0, type: 0x10}}, manufacturerOptions.lumi);
+            return {state: {do_not_disturb: value}};
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_type: {
+        key: ['switch_type'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'toggle': 1, 'momentary': 2};
+            assertString(value, key);
+            value = value.toLowerCase();
+            await entity.write('manuSpecificLumi', {0x000A: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {switch_type: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x000A], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_power_outage_memory: {
+        key: ['power_outage_memory'],
+        convertSet: async (entity, key, value, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['SP-EUC01', 'ZNCZ04LM', 'ZNCZ15LM', 'QBCZ14LM', 'QBCZ15LM', 'SSM-U01', 'SSM-U02', 'DLKZMK11LM', 'DLKZMK12LM',
+                'WS-EUK01', 'WS-EUK02', 'WS-EUK03', 'WS-EUK04', 'QBKG19LM', 'QBKG18LM', 'QBKG20LM', 'QBKG25LM', 'QBKG26LM', 'QBKG28LM', 'QBKG29LM',
+                'QBKG30LM', 'QBKG31LM', 'QBKG32LM', 'QBKG34LM', 'QBKG38LM', 'QBKG39LM', 'QBKG40LM', 'QBKG41LM', 'ZNDDMK11LM', 'ZNLDP13LM',
+                'ZNQBKG31LM', 'WS-USC02', 'WS-USC03', 'WS-USC04', 'ZNQBKG24LM', 'ZNQBKG25LM', 'JWDL001A', 'SSWQD02LM', 'SSWQD03LM',
+                'XDD11LM', 'XDD12LM', 'XDD13LM', 'ZNLDP12LM', 'ZNLDP13LM', 'ZNXDD01LM', 'WS-USC01',
+            ].includes(meta.mapped.model)) {
+                await entity.write('manuSpecificLumi', {0x0201: {value: value ? 1 : 0, type: 0x10}}, manufacturerOptions.lumi);
+            } else if (['ZNCZ02LM', 'QBCZ11LM', 'LLKZMK11LM'].includes(meta.mapped.model)) {
+                const payload = value ?
+                    [[0xaa, 0x80, 0x05, 0xd1, 0x47, 0x07, 0x01, 0x10, 0x01], [0xaa, 0x80, 0x03, 0xd3, 0x07, 0x08, 0x01]] :
+                    [[0xaa, 0x80, 0x05, 0xd1, 0x47, 0x09, 0x01, 0x10, 0x00], [0xaa, 0x80, 0x03, 0xd3, 0x07, 0x0a, 0x01]];
+
+                await entity.write('genBasic', {0xFFF0: {value: payload[0], type: 0x41}}, manufacturerOptions.lumi);
+                await entity.write('genBasic', {0xFFF0: {value: payload[1], type: 0x41}}, manufacturerOptions.lumi);
+            } else if (['ZNCZ11LM', 'ZNCZ12LM'].includes(meta.mapped.model)) {
+                const payload = value ?
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x00, 0x01, 0x10, 0x01] :
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x01, 0x01, 0x10, 0x00];
+
+                await entity.write('genBasic', {0xFFF0: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+            } else {
+                throw new Error('Not supported');
+            }
+            return {state: {power_outage_memory: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['SP-EUC01', 'ZNCZ04LM', 'ZNCZ15LM', 'QBCZ14LM', 'QBCZ15LM', 'SSM-U01', 'SSM-U02', 'DLKZMK11LM', 'DLKZMK12LM',
+                'WS-EUK01', 'WS-EUK02', 'WS-EUK03', 'WS-EUK04', 'QBKG19LM', 'QBKG18LM', 'QBKG20LM', 'QBKG25LM', 'QBKG26LM', 'QBKG28LM', 'QBKG29LM',
+                'QBKG30LM', 'QBKG31LM', 'QBKG32LM', 'QBKG34LM', 'QBKG38LM', 'QBKG39LM', 'QBKG40LM', 'QBKG41LM', 'ZNDDMK11LM', 'ZNLDP13LM',
+                'ZNQBKG31LM', 'WS-USC02', 'WS-USC03', 'WS-USC04', 'ZNQBKG24LM', 'ZNQBKG25LM', 'JWDL001A', 'SSWQD02LM', 'SSWQD03LM',
+                'XDD11LM', 'XDD12LM', 'XDD13LM', 'ZNLDP12LM', 'ZNLDP13LM', 'ZNXDD01LM', 'WS-USC01',
+            ].includes(meta.mapped.model)) {
+                await entity.read('manuSpecificLumi', [0x0201]);
+            } else if (['ZNCZ02LM', 'QBCZ11LM', 'ZNCZ11LM', 'ZNCZ12LM'].includes(meta.mapped.model)) {
+                await entity.read('manuSpecificLumi', [0xFFF0]);
+            } else {
+                throw new Error('Not supported');
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_light_power_outage_memory: {
+        key: ['power_outage_memory'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('genBasic', {0xFF19: {value: value ? 1 : 0, type: 0x10}}, manufacturerOptions.lumi);
+            return {state: {power_outage_memory: value}};
+        },
+    } satisfies Tz.Converter,
+    lumi_power: {
+        key: ['power'],
+        convertGet: async (entity, key, meta) => {
+            const endpoint = meta.device.endpoints.find((e) => e.supportsInputCluster('genAnalogInput'));
+            await endpoint.read('genAnalogInput', ['presentValue']);
+        },
+    } satisfies Tz.Converter,
+    lumi_auto_off: {
+        key: ['auto_off'],
+        convertSet: async (entity, key, value, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['ZNCZ04LM', 'ZNCZ12LM', 'SP-EUC01'].includes(meta.mapped.model)) {
+                await entity.write('manuSpecificLumi', {0x0202: {value: value ? 1 : 0, type: 0x10}}, manufacturerOptions.lumi);
+            } else if (['ZNCZ11LM'].includes(meta.mapped.model)) {
+                const payload = value ?
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x00, 0x02, 0x10, 0x01] :
+                    [0xaa, 0x80, 0x05, 0xd1, 0x47, 0x01, 0x02, 0x10, 0x00];
+
+                await entity.write('genBasic', {0xFFF0: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+            } else {
+                throw new Error('Not supported');
+            }
+            return {state: {auto_off: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (['ZNCZ04LM', 'ZNCZ12LM', 'SP-EUC01'].includes(meta.mapped.model)) {
+                await entity.read('manuSpecificLumi', [0x0202], manufacturerOptions.lumi);
+            } else {
+                throw new Error('Not supported');
+            }
+        },
+    } satisfies Tz.Converter,
+    GZCGQ11LM_detection_period: {
+        key: ['detection_period'],
+        convertSet: async (entity, key, value, meta) => {
+            assertNumber(value, key);
+            value *= 1;
+            await entity.write('manuSpecificLumi', {0x0000: {value: [value], type: 0x21}}, manufacturerOptions.lumi);
+            return {state: {detection_period: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0000], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_curtain_options: {
+        key: ['options'],
+        convertSet: async (entity, key, value, meta) => {
+            assertObject(value);
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            const opts = {
+                reverse_direction: false,
+                hand_open: true,
+                reset_limits: false,
+                ...value,
+            };
+
+            // Legacy names
+            if (value.hasOwnProperty('auto_close')) opts.hand_open = value.auto_close;
+            if (value.hasOwnProperty('reset_move')) opts.reset_limits = value.reset_move;
+
+            if (meta.mapped.model === 'ZNCLDJ12LM' || meta.mapped.model === 'ZNCLDJ14LM') {
+                await entity.write('genBasic', {0xff28: {value: opts.reverse_direction, type: 0x10}}, manufacturerOptions.lumi);
+                await entity.write('genBasic', {0xff29: {value: !opts.hand_open, type: 0x10}}, manufacturerOptions.lumi);
+
+                if (opts.reset_limits) {
+                    await entity.write('genBasic', {0xff27: {value: 0x00, type: 0x10}}, manufacturerOptions.lumi);
+                }
+            } else if (meta.mapped.model === 'ZNCLDJ11LM') {
+                const payload = [
+                    0x07, 0x00, opts.reset_limits ? 0x01 : 0x02, 0x00, opts.reverse_direction ? 0x01 : 0x00, 0x04,
+                    !opts.hand_open ? 0x01 : 0x00, 0x12,
+                ];
+
+                await entity.write('genBasic', {0x0401: {value: payload, type: 0x42}}, manufacturerOptions.lumi);
+
+                // hand_open requires a separate request with slightly different payload
+                payload[2] = 0x08;
+                await entity.write('genBasic', {0x0401: {value: payload, type: 0x42}}, manufacturerOptions.lumi);
+            } else {
+                throw new Error(`lumi_curtain_options set called for not supported model: ${meta.mapped.model}`);
+            }
+
+            // Reset limits is an action, not a state.
+            delete opts.reset_limits;
+            return {state: {options: opts}};
+        },
+        convertGet: async (entity, key, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (meta.mapped.model === 'ZNCLDJ11LM') {
+                await entity.read('genBasic', [0x0401], manufacturerOptions.lumi);
+            } else {
+                throw new Error(`lumi_curtain_options get called for not supported model: ${meta.mapped.model}`);
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_curtain_position_state: {
+        key: ['state', 'position'],
+        options: [exposes.options.invert_cover()],
+        convertSet: async (entity, key, value, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            if (key === 'state' && typeof value === 'string' && value.toLowerCase() === 'stop') {
+                if (meta.mapped.model == 'ZNJLBL01LM') {
+                    const payload = {'presentValue': 2};
+                    await entity.write('genMultistateOutput', payload);
+                } else {
+                    await entity.command('closuresWindowCovering', 'stop', {}, getOptions(meta.mapped, entity));
+                }
+
+                if (!['ZNCLDJ11LM', 'ZNJLBL01LM', 'ZNCLBL01LM'].includes(meta.mapped.model)) {
+                    // The code below is originally added for ZNCLDJ11LM (Koenkk/zigbee2mqtt#4585).
+                    // However, in Koenkk/zigbee-herdsman-converters#4039 it was replaced by reading
+                    // directly from currentPositionLiftPercentage, so that device is excluded.
+                    // For ZNJLBL01LM, in Koenkk/zigbee-herdsman-converters#4163 the position is read
+                    // through onEvent each time the motor stops, so it becomes redundant, and the
+                    // device is excluded.
+                    // The code is left here to avoid breaking compatibility, ideally all devices using
+                    // this converter should be tested so the code can be adjusted/deleted.
+
+                    // Lumi curtain does not send position update on stop, request this.
+                    await entity.read('genAnalogOutput', [0x0055]);
+                }
+
+                return {state: {state: 'STOP'}};
+            } else {
+                const lookup = {'open': 100, 'close': 0, 'on': 100, 'off': 0};
+
+                value = typeof value === 'string' ? value.toLowerCase() : value;
+                if (isString(value)) {
+                    value = getFromLookup(value, lookup);
+                }
+                assertNumber(value);
+                value = meta.options.invert_cover ? 100 - value : value;
+
+                if (['ZNCLBL01LM'].includes(meta.mapped.model)) {
+                    await entity.command('closuresWindowCovering', 'goToLiftPercentage', {percentageliftvalue: value},
+                        getOptions(meta.mapped, entity));
+                } else {
+                    const payload = {0x0055: {value, type: 0x39}};
+                    await entity.write('genAnalogOutput', payload);
+                }
+            }
+        },
+        convertGet: async (entity, key, meta) => {
+            if (!Array.isArray(meta.mapped) && ['ZNCLBL01LM'].includes(meta.mapped.model)) {
+                await entity.read('closuresWindowCovering', ['currentPositionLiftPercentage']);
+            } else {
+                await entity.read('genAnalogOutput', [0x0055]);
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_curtain_battery_voltage: {
+        key: ['voltage'],
+        convertGet: async (entity, key, meta) => {
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            switch (meta.mapped.model) {
+            case 'ZNCLBL01LM':
+                await entity.read('manuSpecificLumi', [0x040B], manufacturerOptions.lumi);
+                break;
+            default:
+                throw new Error(`lumi_curtain_battery_voltage - unsupported model: ${meta.mapped.model}`);
+            }
+        },
+    } satisfies Tz.Converter,
+    lumi_curtain_acn002_charging_status: {
+        key: ['charging_status'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0409], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_curtain_acn002_battery: {
+        key: ['battery'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x040a], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_motion_sensitivity: {
+        key: ['motion_sensitivity'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'low': 1, 'medium': 2, 'high': 3};
+            assertString(value, key);
+            value = value.toLowerCase();
+            await entity.write('manuSpecificLumi', {0x010c: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {motion_sensitivity: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x010c], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_presence: {
+        key: ['presence'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0142], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_monitoring_mode: {
+        key: ['monitoring_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            value = value.toLowerCase();
+            const lookup = {'undirected': 0, 'left_right': 1};
+            await entity.write('manuSpecificLumi', {0x0144: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {monitoring_mode: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0144], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_approach_distance: {
+        key: ['approach_distance'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            value = value.toLowerCase();
+            const lookup = {'far': 0, 'medium': 1, 'near': 2};
+            await entity.write('manuSpecificLumi', {0x0146: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {approach_distance: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0146], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_reset_nopresence_status: {
+        key: ['reset_nopresence_status'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x0157: {value: 1, type: 0x20}}, manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    JTQJBF01LMBW_JTYJGD01LMBW_sensitivity: {
+        key: ['sensitivity'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            value = value.toLowerCase();
+            const lookup = {'low': 0x04010000, 'medium': 0x04020000, 'high': 0x04030000};
+
+            // Timeout of 30 seconds + required (https://github.com/Koenkk/zigbee2mqtt/issues/2287)
+            const options = {...manufacturerOptions.lumi, timeout: 35000};
+            await entity.write('ssIasZone', {0xFFF1: {value: getFromLookup(value, lookup), type: 0x23}}, options);
+            return {state: {sensitivity: value}};
+        },
+    } satisfies Tz.Converter,
+    JTQJBF01LMBW_JTYJGD01LMBW_selfest: {
+        key: ['selftest'],
+        convertSet: async (entity, key, value, meta) => {
+            // Timeout of 30 seconds + required (https://github.com/Koenkk/zigbee2mqtt/issues/2287)
+            const options = {...manufacturerOptions.lumi, timeout: 35000};
+            await entity.write('ssIasZone', {0xFFF1: {value: 0x03010000, type: 0x23}}, options);
+        },
+    } satisfies Tz.Converter,
+    lumi_alarm: {
+        key: ['gas', 'smoke'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x013a], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_density: {
+        key: ['gas_density', 'smoke_density', 'smoke_density_dbm'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x013b], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    JTBZ01AQA_gas_sensitivity: {
+        key: ['gas_sensitivity'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            value = value.toUpperCase();
+            const lookup = {'15%LEL': 1, '10%LEL': 2};
+            await entity.write('manuSpecificLumi', {0x010c: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {gas_sensitivity: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x010c], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_selftest: {
+        key: ['selftest'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumiicLumi', {0x0127: {value: true, type: 0x10}}, manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_buzzer: {
+        key: ['buzzer'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            if (Array.isArray(meta.mapped)) throw new Error(`Not supported for groups`);
+            const attribute = ['JY-GZ-01AQ'].includes(meta.mapped.model) ? 0x013e : 0x013f;
+            value = (value.toLowerCase() === 'alarm') ? 15361 : 15360;
+            await entity.write('manuSpecificLumi', {[`${attribute}`]: {value: [`${value}`], type: 0x23}}, manufacturerOptions.lumi);
+            value = (value === 15361) ? 0 : 1;
+            await entity.write('manuSpecificLumi', {0x0126: {value: [`${value}`], type: 0x20}}, manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_buzzer_manual: {
+        key: ['buzzer_manual_alarm', 'buzzer_manual_mute'],
+        convertGet: async (entity, key, meta) => {
+            if (key === 'buzzer_manual_mute') {
+                await entity.read('manuSpecificLumi', [0x0126], manufacturerOptions.lumi);
+            } else if (key === 'buzzer_manual_alarm') {
+                await entity.read('manuSpecificLumi', [0x013d], manufacturerOptions.lumi);
+            }
+        },
+    } satisfies Tz.Converter,
+    JYGZ01AQ_heartbeat_indicator: {
+        key: ['heartbeat_indicator'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x013c: {value: value ? 1 : 0, type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {heartbeat_indicator: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x013c], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_linkage_alarm: {
+        key: ['linkage_alarm'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x014b: {value: value ? 1 : 0, type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {linkage_alarm: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumiicLumi', [0x014b], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    JTBZ01AQA_state: {
+        key: ['state'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumiicLumi', [0x0139], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_power_outage_count: {
+        key: ['power_outage_count'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0002], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    lumi_trigger_indicator: {
+        key: ['trigger_indicator'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x0152: {value: value ? 1 : 0, type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {trigger_indicator: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0152], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    LLKZMK11LM_interlock: {
+        key: ['interlock'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('genBinaryOutput', {0xff06: {value: value ? 0x01 : 0x00, type: 0x10}}, manufacturerOptions.lumi);
+            return {state: {interlock: value}};
+        },
+    } satisfies Tz.Converter,
+    DJT11LM_vibration_sensitivity: {
+        key: ['sensitivity'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value, key);
+            value = value.toLowerCase();
+            const lookup = {'low': 0x15, 'medium': 0x0B, 'high': 0x01};
+
+            const options = {...manufacturerOptions.lumi, timeout: 35000};
+            await entity.write('genBasic', {0xFF0D: {value: getFromLookup(value, lookup), type: 0x20}}, options);
+            return {state: {sensitivity: value}};
+        },
+    } satisfies Tz.Converter,
+    ZNCLBL01LM_hooks_lock: {
+        key: ['hooks_lock'],
+        convertSet: async (entity, key, value, meta) => {
+            const lookup = {'UNLOCK': 0, 'LOCK': 1};
+            await entity.write('manuSpecificLumi', {0x0427: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {[key]: value}};
+        },
+    } satisfies Tz.Converter,
+    ZNCLBL01LM_hooks_state: {
+        key: ['hooks_state'],
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0428], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    ZNCLBL01LM_hand_open: {
+        key: ['hand_open'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi', {0x0401: {value: !value, type: 0x10}}, manufacturerOptions.lumi);
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x0401], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+    ZNCLBL01LM_limits_calibration: {
+        key: ['limits_calibration'],
+        convertSet: async (entity, key, value, meta) => {
+            switch (value) {
+            case 'start':
+                await entity.write('manuSpecificLumi', {0x0407: {value: 0x01, type: 0x20}}, manufacturerOptions.lumi);
+                break;
+            case 'end':
+                await entity.write('manuSpecificLumi', {0x0407: {value: 0x02, type: 0x20}}, manufacturerOptions.lumi);
+                break;
+            case 'reset':
+                await entity.write('manuSpecificLumi', {0x0407: {value: 0x00, type: 0x20}}, manufacturerOptions.lumi);
+                // also? await entity.write('manuSpecificLumi', {0x0402: {value: 0x00, type: 0x10}}, manufacturerOptions.lumi);
+                break;
+            }
+        },
+    } satisfies Tz.Converter,
+    ZNCJMB14LM: {
+        key: ['theme',
+            'standby_enabled',
+            'beep_volume',
+            'lcd_brightness',
+            'language',
+            'screen_saver_style',
+            'standby_time',
+            'font_size',
+            'lcd_auto_brightness_enabled',
+            'homepage',
+            'screen_saver_enabled',
+            'standby_lcd_brightness',
+            'available_switches',
+            'switch_1_text_icon',
+            'switch_2_text_icon',
+            'switch_3_text_icon',
+        ],
+        convertSet: async (entity, key, value, meta) => {
+            if (key === 'theme') {
+                const lookup = {'classic': 0, 'concise': 1};
+                await entity.write('manuSpecificLumi', {0x0215: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {theme: value}};
+            } else if (key === 'standby_enabled') {
+                await entity.write('manuSpecificLumi', {0x0213: {value: value, type: 0x10}}, manufacturerOptions.lumi);
+                return {state: {standby_enabled: value}};
+            } else if (key === 'beep_volume') {
+                const lookup = {'mute': 0, 'low': 1, 'medium': 2, 'high': 3};
+                await entity.write('manuSpecificLumi', {0x0212: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {beep_volume: value}};
+            } else if (key === 'lcd_brightness') {
+                await entity.write('manuSpecificLumi', {0x0211: {value: value, type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {lcd_brightness: value}};
+            } else if (key === 'language') {
+                const lookup = {'chinese': 0, 'english': 1};
+                await entity.write('manuSpecificLumi', {0x0210: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {language: value}};
+            } else if (key === 'screen_saver_style') {
+                const lookup = {'classic': 1, 'analog clock': 2};
+                await entity.write('manuSpecificLumi', {0x0214: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {screen_saver_style: value}};
+            } else if (key === 'standby_time') {
+                await entity.write('manuSpecificLumi', {0x0216: {value: value, type: 0x23}}, manufacturerOptions.lumi);
+                return {state: {standby_time: value}};
+            } else if (key === 'font_size') {
+                const lookup = {'small': 3, 'medium': 4, 'large': 5};
+                await entity.write('manuSpecificLumi', {0x0217: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {font_size: value}};
+            } else if (key === 'lcd_auto_brightness_enabled') {
+                await entity.write('manuSpecificLumi', {0x0218: {value: value, type: 0x10}}, manufacturerOptions.lumi);
+                return {state: {lcd_auto_brightness_enabled: value}};
+            } else if (key === 'homepage') {
+                const lookup = {'scene': 0, 'feel': 1, 'thermostat': 2, 'switch': 3};
+                await entity.write('manuSpecificLumi', {0x0219: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {homepage: value}};
+            } else if (key === 'screen_saver_enabled') {
+                await entity.write('manuSpecificLumi', {0x0221: {value: value, type: 0x10}}, manufacturerOptions.lumi);
+                return {state: {screen_saver_enabled: value}};
+            } else if (key === 'standby_lcd_brightness') {
+                await entity.write('manuSpecificLumi', {0x0222: {value: value, type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {standby_lcd_brightness: value}};
+            } else if (key === 'available_switches') {
+                const lookup = {'none': 0, '1': 1, '2': 2, '1 and 2': 3, '3': 4, '1 and 3': 5, '2 and 3': 6, 'all': 7};
+                await entity.write('manuSpecificLumi', {0x022b: {value: getFromLookup(value, lookup), type: 0x20}}, manufacturerOptions.lumi);
+                return {state: {available_switches: value}};
+            } else if (key === 'switch_1_text_icon') {
+                const lookup = {'1': 1, '2': 2, '3': 3, '4': 4, '5': 5, '6': 6, '7': 7, '8': 8, '9': 9, '10': 10, '11': 11};
+                const payload = [];
+                const statearr: KeyValue = {};
+                assertObject(value);
+                if (value.hasOwnProperty('switch_1_icon')) {
+                    payload.push(getFromLookup(value.switch_1_icon, lookup));
+                    statearr.switch_1_icon = value.switch_1_icon;
+                } else {
+                    payload.push(1);
+                    statearr.switch_1_icon = '1';
+                }
+                if (value.hasOwnProperty('switch_1_text')) {
+                    payload.push(...value.switch_1_text.split('').map((c: string) => c.charCodeAt(0)));
+                    statearr.switch_1_text = value.switch_1_text;
+                } else {
+                    // @ts-expect-error
+                    payload.push(...''.text.split('').map((c) => c.charCodeAt(0)));
+                    statearr.switch_1_text = '';
+                }
+                await entity.write('manuSpecificLumi', {0x0223: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+                return {state: statearr};
+            } else if (key === 'switch_2_text_icon') {
+                const lookup = {'1': 1, '2': 2, '3': 3, '4': 4, '5': 5, '6': 6, '7': 7, '8': 8, '9': 9, '10': 10, '11': 11};
+                const payload = [];
+                const statearr: KeyValue = {};
+                assertObject(value);
+                if (value.hasOwnProperty('switch_2_icon')) {
+                    payload.push(getFromLookup(value.switch_2_icon, lookup));
+                    statearr.switch_2_icon = value.switch_2_icon;
+                } else {
+                    payload.push(1);
+                    statearr.switch_2_icon = '1';
+                }
+                if (value.hasOwnProperty('switch_2_text')) {
+                    payload.push(...value.switch_2_text.split('').map((c: string) => c.charCodeAt(0)));
+                    statearr.switch_2_text = value.switch_2_text;
+                } else {
+                    // @ts-expect-error
+                    payload.push(...''.text.split('').map((c) => c.charCodeAt(0)));
+                    statearr.switch_2_text = '';
+                }
+                await entity.write('manuSpecificLumi', {0x0224: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+                return {state: statearr};
+            } else if (key === 'switch_3_text_icon') {
+                const lookup = {'1': 1, '2': 2, '3': 3, '4': 4, '5': 5, '6': 6, '7': 7, '8': 8, '9': 9, '10': 10, '11': 11};
+                const payload = [];
+                const statearr: KeyValue = {};
+                assertObject(value);
+                if (value.hasOwnProperty('switch_3_icon')) {
+                    payload.push(getFromLookup(value.switch_3_icon, lookup));
+                    statearr.switch_3_icon = value.switch_3_icon;
+                } else {
+                    payload.push(1);
+                    statearr.switch_3_icon = '1';
+                }
+                if (value.hasOwnProperty('switch_3_text')) {
+                    payload.push(...value.switch_3_text.split('').map((c: string) => c.charCodeAt(0)));
+                    statearr.switch_3_text = value.switch_3_text;
+                } else {
+                    // @ts-expect-error
+                    payload.push(...''.text.split('').map((c) => c.charCodeAt(0)));
+                    statearr.switch_3_text = '';
+                }
+                await entity.write('manuSpecificLumi', {0x0225: {value: payload, type: 0x41}}, manufacturerOptions.lumi);
+                return {state: statearr};
+            } else {
+                throw new Error(`Not supported: '${key}'`);
+            }
+        },
+    } satisfies Tz.Converter,
+    aqara_opple_operation_mode: {
+        key: ['operation_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            assertString(value);
+            // modes:
+            // 0 - 'command' mode. keys send commands. useful for binding
+            // 1 - 'event' mode. keys send events. useful for handling
+            const lookup = {command: 0, event: 1};
+            const endpoint = meta.device.getEndpoint(1);
+            await endpoint.write('manuSpecificLumi', {'mode': getFromLookup(value.toLowerCase(), lookup)},
+                {manufacturerCode: manufacturerOptions.lumi.manufacturerCode});
+            return {state: {operation_mode: value.toLowerCase()}};
+        },
+        convertGet: async (entity, key, meta) => {
+            const endpoint = meta.device.getEndpoint(1);
+            await endpoint.read('manuSpecificLumi', ['mode'], {manufacturerCode: manufacturerOptions.lumi.manufacturerCode});
+        },
+    } satisfies Tz.Converter,
+    lumi_switch_click_mode: {
+        key: ['click_mode'],
+        convertSet: async (entity, key, value, meta) => {
+            await entity.write('manuSpecificLumi',
+                {0x0125: {value: getFromLookup(value, {'fast': 0x1, 'multi': 0x02}), type: 0x20}}, manufacturerOptions.lumi);
+            return {state: {click_mode: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read('manuSpecificLumi', [0x125], manufacturerOptions.lumi);
+        },
+    } satisfies Tz.Converter,
+};
+
+export const legacyFromZigbee = {
+    lumi_click: {
+        cluster: 'genOnOff',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (isLegacyEnabled(options)) {
+                if (model.model === 'WXKG01LM') {
+                    const fromZigbeeStore: KeyValueAny = {};
+                    const deviceID = msg.device.ieeeAddr;
+                    const state = msg.data['onOff'];
+                    const key = `${deviceID}_legacy`;
+                    if (!fromZigbeeStore[key]) {
+                        fromZigbeeStore[key] = {};
+                    }
+                    const current = msg.meta.zclTransactionSequenceNumber;
+                    if (fromZigbeeStore[key].transaction === current) return;
+                    fromZigbeeStore[key].transaction = current;
+
+                    // 0 = click down, 1 = click up, else = multiple clicks
+                    if (state === 0) {
+                        fromZigbeeStore[key].timer = setTimeout(() => {
+                            publish({click: 'long'});
+                            fromZigbeeStore[key].timer = null;
+                            fromZigbeeStore[key].long = Date.now();
+                            fromZigbeeStore[key].long_timer = setTimeout(() => {
+                                fromZigbeeStore[key].long = false;
+                            }, 4000); // After 4000 milliseconds of not receiving long_release we assume it will not happen.
+                            // @ts-expect-error
+                        }, options.long_timeout || 1000); // After 1000 milliseconds of not releasing we assume long click.
+                    } else if (state === 1) {
+                        if (fromZigbeeStore[key].long) {
+                            const duration = Date.now() - fromZigbeeStore[key].long;
+                            publish({click: 'long_release', duration: duration});
+                            fromZigbeeStore[key].long = false;
+                        }
+
+                        if (fromZigbeeStore[key].timer) {
+                            clearTimeout(fromZigbeeStore[key].timer);
+                            fromZigbeeStore[key].timer = null;
+                            publish({click: 'single'});
+                        }
+                    } else {
+                        const clicks = msg.data['32768'];
+                        const actionLookup: KeyValueAny = {1: 'single', 2: 'double', 3: 'triple', 4: 'quadruple'};
+                        const payload = actionLookup[clicks] ? actionLookup[clicks] : 'many';
+                        publish({click: payload});
+                    }
+                }
+                if (['QBKG03LM', 'QBKG12LM', 'QBKG22LM'].includes(model.model)) {
+                    if (!msg.data['61440']) {
+                        const mapping: KeyValueAny = {4: 'left', 5: 'right', 6: 'both'};
+                        const button = mapping[msg.endpoint.ID];
+                        return {click: button};
+                    }
+                }
+                if (['QBKG04LM', 'QBKG11LM', 'QBKG21LM'].includes(model.model)) {
+                    if (!msg.data['61440']) {
+                        return {click: 'single'};
+                    }
+                }
+                if (['WXKG03LM_rev1', 'WXKG03LM_rev2'].includes(model.model)) {
+                    return {click: 'single'};
+                }
+                if (model.model === 'WXKG11LM') {
+                    const data = msg.data;
+                    let clicks;
+
+                    if (data.onOff) {
+                        clicks = 1;
+                    } else if (data['32768']) {
+                        clicks = data['32768'];
+                    }
+
+                    const actionLookup: KeyValueAny = {1: 'single', 2: 'double', 3: 'triple', 4: 'quadruple'};
+                    if (actionLookup[clicks]) {
+                        return {click: actionLookup[clicks]};
+                    }
+                }
+                if (['WXKG02LM_rev1', 'WXKG02LM_rev2'].includes(model.model)) {
+                    const lookup: KeyValueAny = {1: 'left', 2: 'right', 3: 'both'};
+                    return {click: lookup[msg.endpoint.ID]};
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_click_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (model.model === 'QBKG11LM') {
+                if (isLegacyEnabled(options)) {
+                    if ([1, 2].includes(msg.data.presentValue)) {
+                        const times: KeyValueAny = {1: 'single', 2: 'double'};
+                        return {click: times[msg.data.presentValue]};
+                    }
+                }
+            }
+            if (model.model === 'QBKG12LM') {
+                if (isLegacyEnabled(options)) {
+                    if ([1, 2].includes(msg.data.presentValue)) {
+                        const mapping: KeyValueAny = {5: 'left', 6: 'right', 7: 'both'};
+                        const times: KeyValueAny = {1: 'single', 2: 'double'};
+                        const button = mapping[msg.endpoint.ID];
+                        return {click: `${button}_${times[msg.data.presentValue]}`};
+                    }
+                }
+            }
+            if (model.model === 'WXKG02LM_rev2') {
+                const fromZigbeeStore: KeyValueAny = {};
+                // Somestime WXKG02LM sends multiple messages on a single click, this prevents handling
+                // of a message with the same transaction sequence number twice.
+                const current = msg.meta.zclTransactionSequenceNumber;
+                if (fromZigbeeStore[msg.device.ieeeAddr + 'legacy'] === current) return;
+                fromZigbeeStore[msg.device.ieeeAddr + 'legacy'] = current;
+
+                const buttonLookup: KeyValueAny = {1: 'left', 2: 'right', 3: 'both'};
+                const button = buttonLookup[msg.endpoint.ID];
+                const value = msg.data['presentValue'];
+
+                const actionLookup: KeyValueAny = {
+                    0: 'long',
+                    1: null,
+                    2: 'double',
+                };
+
+                const action = actionLookup[value];
+
+                if (button) {
+                    if (isLegacyEnabled(options)) {
+                        return {click: button + (action ? `_${action}` : '')};
+                    }
+                }
+            }
+            if (model.model === 'QBKG25LM') {
+                if (isLegacyEnabled(options)) {
+                    if ([1, 2, 3, 0, 255].includes(msg.data.presentValue)) {
+                        const mapping: KeyValueAny = {41: 'left', 42: 'center', 43: 'right'};
+                        const times: KeyValueAny = {1: 'single', 2: 'double', 3: 'triple', 0: 'hold', 255: 'release'};
+                        const button = mapping[msg.endpoint.ID];
+                        return {action: `${button}_${times[msg.data.presentValue]}`};
+                    }
+                } else {
+                    return fromZigbee.lumi_action_multistate.convert(model, msg, publish, options, meta);
+                }
+            }
+            if (['WXKG11LM', 'WXKG12LM', 'WXKG03LM_rev2'].includes(model.model)) {
+                if (isLegacyEnabled(options)) {
+                    const value = msg.data['presentValue'];
+                    const lookup: KeyValueAny = {
+                        1: {click: 'single'}, // single click
+                        2: {click: 'double'}, // double click
+                    };
+                    return lookup[value] ? lookup[value] : null;
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_buttons: {
+        cluster: 'genOnOff',
+        type: ['attributeReport'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (isLegacyEnabled(options)) {
+                const mapping: KeyValueAny = {4: 'left', 5: 'right'};
+                const button = mapping[msg.endpoint.ID];
+                if (button) {
+                    const payload: KeyValueAny = {};
+                    payload[`button_${button}`] = msg.data['onOff'] === 1 ? 'release' : 'hold';
+                    return payload;
+                }
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_action: {
+        cluster: 'genOnOff',
+        type: ['attributeReport'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            if (isLegacyEnabled(options)) {
+                return {action: getKey(model.endpoint(msg.device), msg.endpoint.ID)};
+            } else {
+                return fromZigbee.lumi_action.convert(model, msg, publish, options, meta);
+            }
+        },
+    } satisfies Fz.Converter,
+    lumi_action_multistate: {
+        cluster: 'genMultistateInput',
+        type: ['attributeReport', 'readResponse'],
+        options: [exposes.options.legacy()],
+        convert: (model, msg, publish, options, meta) => {
+            // refactor to lumi_multistate_action
+            if (isLegacyEnabled(options)) {
+                const button = getKey(model.endpoint(msg.device), msg.endpoint.ID);
+                const value = msg.data['presentValue'];
+                const actionLookup: KeyValueAny = {0: 'long', 1: null, 2: 'double'};
+                const action = actionLookup[value];
+
+                if (button) {
+                    return {action: `${button}${(action ? `_${action}` : '')}`};
+                }
+            } else {
+                return fromZigbee.lumi_action_multistate.convert(model, msg, publish, options, meta);
+            }
+        },
+    } satisfies Fz.Converter,
+};
+
+export const legacyToZigbee = {
+
 };
 
 exports.buffer2DataObject = buffer2DataObject;
